% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{aic.mle}
\alias{aic.mle}
\title{Method for obtaining the AIC of an `mle` object.}
\usage{
\method{aic}{mle}(x)
}
\arguments{
\item{x}{the `mle` object to obtain the AIC of}
}
\value{
Numeric AIC value.
}
\description{
Method for obtaining the AIC of an `mle` object.
}
