% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{aic.mle_fit}
\alias{aic.mle_fit}
\title{Method for obtaining the AIC of an `mle_fit` object.}
\usage{
\method{aic}{mle_fit}(x)
}
\arguments{
\item{x}{the `mle_fit` object to obtain the AIC of}
}
\value{
Numeric AIC value.
}
\description{
Method for obtaining the AIC of an `mle_fit` object.
}
