% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{bias.boot}
\alias{bias.boot}
\title{Computes the estimate of the bias of a \code{boot} object.}
\usage{
\method{bias}{boot}(x, par = NULL, ...)
}
\arguments{
\item{x}{the \code{boot} object to compute the bias of.}

\item{par}{if the true parameter value is known, you may provide it;
otherwise we use the MLE of \code{par}.}

\item{...}{pass additional arguments}
}
\description{
Computes the estimate of the bias of a \code{boot} object.
}
