% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{bias.mle}
\alias{bias.mle}
\title{Computes the bias of an \code{mle} object.}
\usage{
\method{bias}{mle}(x, ...)
}
\arguments{
\item{x}{the \code{mle} object to compute the bias of.}

\item{...}{additional arguments to pass}
}
\description{
Computes the bias of an \code{mle} object.
}
