% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_emp.R
\name{bias.mle_emp}
\alias{bias.mle_emp}
\title{Computes the estimate of the bias of a `mle_emp` object.}
\usage{
\method{bias}{mle_emp}(x, par, ...)
}
\arguments{
\item{x}{the `mle_emp` object to compute the bias of.}

\item{par}{The true parameter value.}

\item{...}{pass additional arguments}
}
\description{
Computes the estimate of the bias of a `mle_emp` object.
}
