% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{bias.mle_exp}
\alias{bias.mle_exp}
\title{Computes the Bias of an Exponential MLE}
\usage{
\method{bias}{mle_exp}(x, par = NULL, ...)
}
\arguments{
\item{x}{an `mle_exp` object from which to compute the bias.}

\item{par}{a numeric value representing the true rate parameter. If NULL, the
point estimate from `x` is used.}

\item{...}{additional arguments (not used).}
}
\value{
Returns a named numeric value representing the estimated bias of the rate parameter.
}
\description{
This function computes the bias of the Maximum Likelihood Estimator (MLE) of
the rate parameter for an Exponential distribution. The bias is
`1/(nobs(x)-1)*par` where `x` is an `mle_exp` object and `par` is the true
rate parameter. If `par` is not provided, the point estimate from `x` is
used to estimate the bias.
}
\examples{
\dontrun{
  obs <- rexp(100, rate = 0.5)
  rate.hat <- mle_exp(obs)
  print(bias(rate.hat, par = 0.5))
}

}
