% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{bias.mle_fit_boot}
\alias{bias.mle_fit_boot}
\title{Computes the estimate of the bias of a `mle_fit_boot` object.}
\usage{
\method{bias}{mle_fit_boot}(x, theta = NULL, ...)
}
\arguments{
\item{x}{the `mle_fit_boot` object to compute the bias of.}

\item{theta}{true parameter value (not used for `mle_fit_boot`).}

\item{...}{pass additional arguments (not used)}
}
\value{
Numeric vector of estimated bias (mean of bootstrap replicates minus
  original estimate).
}
\description{
Computes the estimate of the bias of a `mle_fit_boot` object.
}
\examples{
set.seed(1)
b <- boot::boot(rexp(50, 2), function(d, i) 1/mean(d[i]), R = 99)
fit <- mle_boot(b)
bias(fit)
}
