% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{bias.mle_normal}
\alias{bias.mle_normal}
\title{Computes the bias of an \code{mle_normal} object.}
\usage{
\method{bias}{mle_normal}(x, theta)
}
\arguments{
\item{x}{the \code{mle} object to compute the bias of.}

\item{theta}{the true parameter value}
}
\description{
Computes the bias of an \code{mle_normal} object.
}
