% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal_mu_var.R
\name{bias.mle_normal_mu_var}
\alias{bias.mle_normal_mu_var}
\title{Computes the bias of an `mle_normal_mu_var` object.}
\usage{
\method{bias}{mle_normal_mu_var}(x, par = NULL, ...)
}
\arguments{
\item{x}{the `mle` object to compute the bias of.}

\item{par}{the true parameter value, defaults to unknown (NULL)}

\item{...}{unused additional arguments}
}
\description{
Computes the bias of an `mle_normal_mu_var` object.
}
