% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.mle.R
\name{bias}
\alias{bias}
\title{Generic method for computing the bias of an estimator object.}
\usage{
bias(x, theta = NULL, ...)
}
\arguments{
\item{x}{the object to compute the bias of.}

\item{theta}{true parameter value. usually, this is unknown (NULL), in which
case we estimate the bias}

\item{...}{pass additional arguments}
}
\value{
The bias of the estimator. The return type depends on the specific
        method.
}
\description{
Generic method for computing the bias of an estimator object.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), loglike = -120, nobs = 100L)
bias(fit, theta = c(mu = 5, sigma2 = 4))
}
