% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{bic.mle}
\alias{bic.mle}
\title{Method for obtaining the BIC of an `mle` object.}
\usage{
\method{bic}{mle}(x)
}
\arguments{
\item{x}{the `mle` object to obtain the BIC of}
}
\value{
Numeric BIC value, or \code{NA_real_} if log-likelihood or
  observation count is unavailable.
}
\description{
Method for obtaining the BIC of an `mle` object.
}
