% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{block_diag}
\alias{block_diag}
\title{Build a block-diagonal matrix from a list of square matrices.}
\usage{
block_diag(blocks, dims)
}
\arguments{
\item{blocks}{List of matrices (scalars are promoted to 1x1 matrices).}

\item{dims}{Integer vector of block dimensions.}
}
\value{
A square matrix with blocks placed along the diagonal.
}
\description{
Build a block-diagonal matrix from a list of square matrices.
}
\keyword{internal}
