% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_estimate}
\alias{boot_estimate}
\title{Bootstrap method}
\usage{
boot_estimate(xs, stat, B = 1000, ...)
}
\arguments{
\item{xs}{data}

\item{stat}{statistic of data}

\item{B}{bootstrap replicates}

\item{...}{additional arguments to pass.}
}
\description{
Bootstrap method
}
