% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{cdf.mle}
\alias{cdf.mle}
\title{CDF of the asymptotic distribution of an MLE.}
\usage{
\method{cdf}{mle}(x, ...)
}
\arguments{
\item{x}{An \code{mle} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function computing the CDF at given points.
}
\description{
CDF of the asymptotic distribution of an MLE.
}
