% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{coef.mle}
\alias{coef.mle}
\title{Extract coefficients from an `mle` object.}
\usage{
\method{coef}{mle}(object, ...)
}
\arguments{
\item{object}{the `mle` object}

\item{...}{additional arguments (not used)}
}
\value{
Named numeric vector of parameter estimates.
}
\description{
Provides compatibility with the standard R \code{coef()} generic.
Returns the same values as \code{params(x)}.
}
