% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{collect_or_null}
\alias{collect_or_null}
\title{Collect a field from a list of MLEs, returning NULL if any are NULL.}
\usage{
collect_or_null(mles, extractor)
}
\arguments{
\item{mles}{List of mle_fit objects.}

\item{extractor}{Function that extracts the field from a single MLE.}
}
\value{
A list of field values, or NULL if any value is NULL.
}
\description{
Collect a field from a list of MLEs, returning NULL if any are NULL.
}
\keyword{internal}
