% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\alias{combine.list}
\alias{combine.mle}
\title{Combine independent MLEs for the same parameter.}
\usage{
combine(x, ...)

\method{combine}{list}(x, ...)

\method{combine}{mle}(x, ...)
}
\arguments{
\item{x}{An \code{mle} object, or a list of \code{mle} objects.}

\item{...}{Additional \code{mle} objects to combine.}
}
\value{
An \code{mle} object representing the optimally weighted combination.
}
\description{
Given multiple independent MLEs that estimate the same parameter \eqn{\theta},
produces an optimally weighted combination using inverse-variance (Fisher
information) weighting.
}
\details{
The combined estimator has:
\itemize{
  \item \code{theta.hat}: \eqn{(\sum I_i)^{-1} \sum I_i \hat\theta_i}
  \item \code{sigma}: \eqn{(\sum I_i)^{-1}}
  \item \code{info}: \eqn{\sum I_i}
  \item \code{nobs}: sum of individual sample sizes
}

When the Fisher information matrix is not directly available but the
variance-covariance matrix is, the FIM is computed as \code{ginv(vcov)}.

For the legacy interface that accepts a list, see \code{\link{mle_weighted}}.
}
\examples{
# Three independent estimates of the same rate
fit1 <- mle(theta.hat = c(lambda = 2.1), sigma = matrix(0.04), nobs = 50L)
fit2 <- mle(theta.hat = c(lambda = 1.9), sigma = matrix(0.02), nobs = 100L)
fit3 <- mle(theta.hat = c(lambda = 2.0), sigma = matrix(0.03), nobs = 70L)

comb <- combine(fit1, fit2, fit3)
params(comb)
se(comb)
}
\seealso{
\code{\link{mle_weighted}}, \code{\link{joint}}
}
