% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{confint.boot}
\alias{confint.boot}
\title{Function to compute the confidence intervals of \code{mle} objects.}
\usage{
\method{confint}{boot}(object, parm = NULL, level = 0.95, type = "perc", ...)
}
\arguments{
\item{object}{the \code{boot} object to compute the confidence intervals for}

\item{parm}{parameter indexes to compute the confidence intervals for,
defaults to all. NOTE: not implemented so ignored for now.}

\item{level}{confidence level, defaults to 0.95 (alpha=.05)}

\item{type}{A vector of character strings representing the type of intervals
required. The value should be any subset of the values
\code{c("norm","basic", "stud", "perc", "bca")} or simply "all".}

\item{...}{additional arguments to pass}
}
\description{
Function to compute the confidence intervals of \code{mle} objects.
}
