% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{confint.mle}
\alias{confint.mle}
\title{Function to compute the confidence intervals of `mle` objects.}
\usage{
\method{confint}{mle}(object, parm = NULL, level = 0.95, use_t_dist = FALSE, ...)
}
\arguments{
\item{object}{the `mle` object to compute the confidence intervals for}

\item{parm}{the parameters to compute the confidence intervals for (not used)}

\item{level}{confidence level, defaults to 0.95 (alpha=.05)}

\item{use_t_dist}{logical, whether to use the t-distribution to compute
the confidence intervals.}

\item{...}{additional arguments to pass}
}
\description{
Function to compute the confidence intervals of `mle` objects.
}
