% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{confint.mle_boot}
\alias{confint.mle_boot}
\title{Method for obtained the confidence interval of an `mle_boot` object.}
\usage{
\method{confint}{mle_boot}(x, level = 0.95, ...)
}
\arguments{
\item{x}{the `mle_boot` object to obtain the confidence interval of}

\item{level}{the confidence level}

\item{...}{additional arguments to pass}
}
\description{
Method for obtained the confidence interval of an `mle_boot` object.
}
