% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_emp.R
\name{confint.mle_emp}
\alias{confint.mle_emp}
\title{Method for obtained the confidence interval of an `mle_emp` object.}
\usage{
\method{confint}{mle_emp}(x, level = 0.95, ...)
}
\arguments{
\item{x}{the `mle_emp` object to obtain the confidence interval of}

\item{level}{the confidence level}

\item{...}{additional arguments to pass}
}
\description{
Method for obtained the confidence interval of an `mle_emp` object.
}
