% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{confint.mle_fit}
\alias{confint.mle_fit}
\title{Function to compute the confidence intervals of `mle_fit` objects.}
\usage{
\method{confint}{mle_fit}(object, parm = NULL, level = 0.95, use_t_dist = FALSE, ...)
}
\arguments{
\item{object}{the `mle_fit` object to compute the confidence intervals for}

\item{parm}{character or integer vector of parameters to return intervals for.
If NULL (default), all parameters are returned.}

\item{level}{confidence level, defaults to 0.95 (alpha=.05)}

\item{use_t_dist}{logical, whether to use the t-distribution to compute
the confidence intervals.}

\item{...}{additional arguments to pass}
}
\value{
Matrix of confidence intervals with columns for lower and upper bounds.
}
\description{
Function to compute the confidence intervals of `mle_fit` objects.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), loglike = -120, nobs = 100L)
confint(fit)
}
