% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{confint}
\alias{confint}
\title{Compute the confidence interval given an estimator object `x`.}
\usage{
confint(x, level = 0.95, ...)
}
\arguments{
\item{x}{the estimator object}

\item{level}{confidence interval level}

\item{...}{additional arguments to pass}
}
\description{
Compute the confidence interval given an estimator object `x`.
}
