% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{cramer.test}
\alias{cramer.test}
\title{Cramer Von Mises hypothesis test}
\usage{
cramer.test(obs, ref)
}
\arguments{
\item{obs}{a sample of observations}

\item{ref}{a sample from the reference distribution}
}
\value{
hypothesis test object
}
\description{
Cramer Von Mises hypothesis test
}
