% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{density.mle_boot}
\alias{density.mle_boot}
\title{PDF of the empirical distribution of bootstrap replicates.}
\usage{
\method{density}{mle_boot}(x, ...)
}
\arguments{
\item{x}{An \code{mle_boot} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function computing the empirical PMF at given points.
}
\description{
PDF of the empirical distribution of bootstrap replicates.
}
