% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{density.mle_fit}
\alias{density.mle_fit}
\title{PDF of the asymptotic distribution of an MLE.}
\usage{
\method{density}{mle_fit}(x, ...)
}
\arguments{
\item{x}{An \code{mle_fit} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function computing the PDF at given points.
}
\description{
Returns a closure computing the probability density function of the
asymptotic normal (univariate) or multivariate normal (multivariate)
distribution of the MLE.
}
