% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{density.mle_fit_boot}
\alias{density.mle_fit_boot}
\title{PDF of the empirical distribution of bootstrap replicates.}
\usage{
\method{density}{mle_fit_boot}(x, ...)
}
\arguments{
\item{x}{An \code{mle_fit_boot} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function computing the empirical PMF at given points.
}
\description{
PDF of the empirical distribution of bootstrap replicates.
}
\examples{
set.seed(1)
b <- boot::boot(rexp(50, 2), function(d, i) 1/mean(d[i]), R = 99)
fit <- mle_boot(b)
density(fit)
}
