% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{dim.mle}
\alias{dim.mle}
\title{Dimension (number of parameters) of an MLE.}
\usage{
\method{dim}{mle}(x)
}
\arguments{
\item{x}{An \code{mle} object.}
}
\value{
Integer; the number of parameters.
}
\description{
Dimension (number of parameters) of an MLE.
}
