% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{dim.mle_boot}
\alias{dim.mle_boot}
\title{Dimension (number of parameters) of a bootstrap MLE.}
\usage{
\method{dim}{mle_boot}(x)
}
\arguments{
\item{x}{An \code{mle_boot} object.}
}
\value{
Integer; the number of parameters.
}
\description{
Dimension (number of parameters) of a bootstrap MLE.
}
