% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{dim.mle_fit_boot}
\alias{dim.mle_fit_boot}
\title{Dimension (number of parameters) of a bootstrap MLE.}
\usage{
\method{dim}{mle_fit_boot}(x)
}
\arguments{
\item{x}{An \code{mle_fit_boot} object.}
}
\value{
Integer; the number of parameters.
}
\description{
Dimension (number of parameters) of a bootstrap MLE.
}
\examples{
set.seed(1)
b <- boot::boot(rexp(50, 2), function(d, i) 1/mean(d[i]), R = 99)
fit <- mle_boot(b)
dim(fit)
}
