% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{distr}
\alias{distr}
\title{Generic function for computing the estimator of \code{f(x)} where \code{x}
models an estimator.}
\usage{
distr(x, f, n = 1000, ...)
}
\arguments{
\item{x}{an estimator object.}

\item{f}{a function of the \code{x} object.}

\item{n}{number of samples to take to estimate \code{f(x)}  with.}

\item{...}{additional arguments to pass.}
}
\description{
Generic function for computing the estimator of \code{f(x)} where \code{x}
models an estimator.
}
