% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{exp_conditional_loglike}
\alias{exp_conditional_loglike}
\title{Log-Likelihood Function Generator for the Exponential Distribution
with a rate parameter that is a function of predictors in `df`,}
\usage{
exp_conditional_loglike(df, resp, rate)
}
\arguments{
\item{df}{a numeric vector representing a sample of observations.}

\item{resp}{a function that returns the response variable given a row
from a data frame.}

\item{rate}{a function that returns the rate given a row from a data frame
and a parameter vector `beta`}
}
\value{
Returns a function that computes the conditional log-likelihood
}
\description{
`resp(df) ~ exp(rate(df, beta))`
}
\details{
where `beta` is a vector of parameters.

The function returned by this function is suitable for use with
`optim` or `nlm` to find the maximum likelihood estimate of `beta`.

Note that the expected value is `1/rate(...)`.
}
