% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{exp_fim}
\alias{exp_fim}
\title{Fisher Information Function Generator for the Exponential Distribution}
\usage{
exp_fim(x)
}
\arguments{
\item{x}{a numeric vector representing a sample of observations.}
}
\value{
Returns a function that computes the Fisher Information of `x` given
        a rate parameter.
}
\description{
This function returns a function of the rate parameter that computes the
Fisher Information of the given data assuming an Exponential distribution.
}
