% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{exp_loglike}
\alias{exp_loglike}
\title{Log-Likelihood Function Generator for the Exponential Distribution}
\usage{
exp_loglike(x)
}
\arguments{
\item{x}{a numeric vector representing a sample of observations.}
}
\value{
Returns a function that computes the log-likelihood of `x` given a
        rate parameter.
}
\description{
This function returns a function of the rate parameter that computes the
log-likelihood of the given data assuming an Exponential distribution.
}
\examples{
\dontrun{
  obs <- rexp(100, rate = 0.5)
  loglike <- exp_loglike(obs)
  plot(loglike, from = 0, to = 2)
}

}
