% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{exp_rate_fisher_info}
\alias{exp_rate_fisher_info}
\title{Fisher information function generator for the exponential
distribution}
\usage{
exp_rate_fisher_info(x)
}
\arguments{
\item{x}{data}
}
\description{
Fisher information function generator for the exponential
distribution
}
