% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp_rate.R
\name{exp_rate_loglike}
\alias{exp_rate_loglike}
\title{log-likelihood function generator given data `x` for the exponential
distribution}
\usage{
exp_rate_loglike(x)
}
\arguments{
\item{x}{data}
}
\description{
log-likelihood function generator given data `x` for the exponential
distribution
}
