% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp_rate.R
\name{exp_rate_score}
\alias{exp_rate_score}
\title{score (derivative of log-likelihood) function generator given data `x`
for the exponential distribution}
\usage{
exp_rate_score(x)
}
\arguments{
\item{x}{data}
}
\description{
score (derivative of log-likelihood) function generator given data `x`
for the exponential distribution
}
