% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{exp_score}
\alias{exp_score}
\title{Score Function Generator for the Exponential Distribution}
\usage{
exp_score(x)
}
\arguments{
\item{x}{a numeric vector representing a sample of observations.}
}
\value{
Returns a function that computes the score of `x` given a rate
        parameter.
}
\description{
This function returns a function of the rate parameter that computes the
score (derivative of log-likelihood) of the given data assuming an
Exponential distribution.
}
