% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{fim.mle}
\alias{fim.mle}
\title{Function for obtaining the fisher information matrix of an `mle` object.}
\usage{
\method{fim}{mle}(x, ...)
}
\arguments{
\item{x}{the `mle` object to obtain the fisher information of.}

\item{...}{pass additional arguments}
}
\description{
Function for obtaining the fisher information matrix of an `mle` object.
}
