% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{fim.mle_boot}
\alias{fim.mle_boot}
\title{Computes the variance-covariance matrix of \code{boot} object.}
\usage{
\method{fim}{mle_boot}(object, ...)
}
\arguments{
\item{object}{the \code{boot} object to obtain the variance-covariance of}

\item{...}{additional arguments to pass}
}
\description{
Computes the variance-covariance matrix of \code{boot} object.
}
