% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{fim_or_ginv_vcov}
\alias{fim_or_ginv_vcov}
\title{Extract FIM from an MLE, falling back to ginv(vcov).}
\usage{
fim_or_ginv_vcov(m)
}
\arguments{
\item{m}{An mle_fit object.}
}
\value{
A matrix (FIM).
}
\description{
Extract FIM from an MLE, falling back to ginv(vcov).
}
\keyword{internal}
