% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{fisher_info.mle}
\alias{fisher_info.mle}
\title{Function for obtaining the fisher information matrix of an \code{mle} object.}
\usage{
\method{fisher_info}{mle}(x, ...)
}
\arguments{
\item{x}{the \code{mle} object to obtain the fisher information of.}

\item{...}{unused by \code{mle} objects. particular specializations of
\code{mle} objects may use it, however.}
}
\description{
Function for obtaining the fisher information matrix of an \code{mle} object.
}
