% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{fisher_info}
\alias{fisher_info}
\title{Generic method for obtaining the fisher information
matrix of an \code{mle} object.}
\usage{
fisher_info(x, ...)
}
\arguments{
\item{x}{the object to obtain the fisher information of}

\item{...}{additional arguments to pass}
}
\description{
Fisher information is a way of measuring the amount of
information that an observable random variable `X`
carries about an unknown parameter \code{theta}
upon which the probability of `X` depends.
}
\details{
The inverse of the Fisher information matrix
is the variance-covariance of the MLE for
\code{theta}.
}
