% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerial_mle.R
\name{fisher_scoring}
\alias{fisher_scoring}
\title{Fisher scoring algorithm.}
\usage{
fisher_scoring(theta0, info, score, eps = 1e-05, max_iter = 250L)
}
\arguments{
\item{theta0}{initial guess of theta with \eqn{p} components}

\item{info}{information matrix function of type \eqn{R^p -> R^{p \times q}}}

\item{score}{score function of type \eqn{R^p -> R^p}}

\item{eps}{stopping condition}

\item{max_iter}{maximum number of iterations}
}
\description{
Fisher scoring algorithm.
}
