% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_mle.R
\name{fn_distr.mle}
\alias{fn_distr.mle}
\title{Computes the distribution of \code{f(x)} as a function of \code{x}, where
\code{x} models a \code{mle} object.}
\usage{
\method{fn_distr}{mle}(x, f, n = 1000, ...)
}
\arguments{
\item{x}{an \code{mle} object.}

\item{f}{a function that accepts objects like x (e.g., a vector).}

\item{n}{number of samples to take from \code{x} to estimate distribution
of \code{f(x)}.}

\item{...}{additional arguments to pass to the \code{mle} sampler.}
}
\description{
Since \code{x} is an MLE, asymptotically,
\code{f(x) ~ N(mean(f(x)),vcov(f(x)))}.
}
