% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{fn_distr}
\alias{fn_distr}
\title{Generic function for computing the distribution of \code{f(x)} where \code{x}
models another distribution object.}
\usage{
fn_distr(x, f, n = 1000, ...)
}
\arguments{
\item{x}{a \code{distr} object}

\item{f}{a function of the \code{x} object}

\item{n}{number of samples to take to estimate \code{f(x)}  with}

\item{...}{additional arguments to pass}
}
\description{
Generic function for computing the distribution of \code{f(x)} where \code{x}
models another distribution object.
}
