% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_scoring.R
\name{gradient_ascent}
\alias{gradient_ascent}
\title{Gradient ascent}
\usage{
gradient_ascent(theta0, loglike, score = NULL, eps = 0.001, max_iter = 250L)
}
\arguments{
\item{theta0}{initial guess of theta with \eqn{p} components}

\item{loglike}{log-likelihood function}

\item{score}{score function of type \eqn{R^p -> R^p}}

\item{eps}{stopping condition}

\item{max_iter}{maximum iterations}
}
\description{
Gradient ascent
}
