% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{inv_cdf.mle}
\alias{inv_cdf.mle}
\title{Quantile function of the asymptotic distribution of an MLE.}
\usage{
\method{inv_cdf}{mle}(x, ...)
}
\arguments{
\item{x}{An \code{mle} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function computing quantiles for given probabilities.
}
\description{
Quantile function of the asymptotic distribution of an MLE.
}
