% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{inv_cdf.mle_fit}
\alias{inv_cdf.mle_fit}
\title{Quantile function of the asymptotic distribution of an MLE.}
\usage{
\method{inv_cdf}{mle_fit}(x, ...)
}
\arguments{
\item{x}{An \code{mle_fit} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A function computing quantiles for given probabilities.
}
\description{
Quantile function of the asymptotic distribution of an MLE.
}
\examples{
fit <- mle(theta.hat = c(mu = 5), sigma = matrix(0.1))
q <- inv_cdf(fit)
q(0.975)
}
