% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_mle.R
\name{is_converged}
\alias{is_converged}
\title{is_converged}
\usage{
is_converged(x)
}
\arguments{
\item{x}{the `mle` object}

\item{...}{additional arguments to pass}
}
\description{
Function to determine whether an `mle` object has converged.
}
