% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{is_mle}
\alias{is_mle}
\title{Determine if an object \code{x} is an \code{mle} object.}
\usage{
is_mle(x)
}
\arguments{
\item{x}{the object to test}
}
\description{
Determine if an object \code{x} is an \code{mle} object.
}
