% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_emp.R
\name{is_mle_emp}
\alias{is_mle_emp}
\title{Determine if an object is an `mle_emp` object.}
\usage{
is_mle_emp(x)
}
\arguments{
\item{x}{the object to test}
}
\description{
Determine if an object is an `mle_emp` object.
}
