% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{loglik_val.mle}
\alias{loglik_val.mle}
\title{Method for obtaining the log-likelihood of an `mle` object.}
\usage{
\method{loglik_val}{mle}(x, ...)
}
\arguments{
\item{x}{the log-likelihood `l` evaluated at `x`, `l(x)`.}

\item{...}{additional arguments to pass}
}
\value{
the log-likelihood of the fitted mle object `x`
}
\description{
Method for obtaining the log-likelihood of an `mle` object.
}
