% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.mle.R
\name{loglik_val}
\alias{loglik_val}
\title{Generic method for obtaining the log-likelihood value of a fitted MLE
object.}
\usage{
loglik_val(x, ...)
}
\arguments{
\item{x}{the object to obtain the log-likelihood of}

\item{...}{additional arguments to pass}
}
\description{
Generic method for obtaining the log-likelihood value of a fitted MLE
object.
}
