% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{loglike}
\alias{loglike}
\title{Generic method for obtaining the log-likelihood of an MLE fit.}
\usage{
loglike(x, ...)
}
\arguments{
\item{x}{the object to obtain the log-likelihood of}

\item{...}{additional arguments to pass}
}
\description{
Generic method for obtaining the log-likelihood of an MLE fit.
}
