% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{make_mle}
\alias{make_mle}
\title{\code{make_mle} makes an \code{mle} object.}
\usage{
make_mle(
  theta.hat,
  loglike = NULL,
  sigma = NULL,
  info = NULL,
  obs = NULL,
  sample_size = NULL
)
}
\arguments{
\item{theta.hat}{the MLE}

\item{loglike}{the log-likelihood of the MLE given the data}

\item{sigma}{the variance-covariance matrix of the MLE}

\item{info}{the variance-covariance matrix of the MLE}

\item{obs}{observation (sample) data}

\item{sample_size}{the variance-covariance matrix of the MLE}
}
\description{
\code{make_mle} makes an \code{mle} object.
}
