% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{marginal.mle}
\alias{marginal.mle}
\title{Method for obtaining the marginal distribution of an MLE
that is based on asymptotic assumptions:}
\usage{
\method{marginal}{mle}(x, indices)
}
\arguments{
\item{x}{The distribution object.}

\item{indices}{The indices of the marginal distribution to obtain.}
}
\value{
An \code{mle} object representing the marginal distribution for the
  selected parameter indices.
}
\description{
`x ~ MVN(params(x), inv(H)(x))`
}
\details{
where H is the (observed or expecation) Fisher information
matrix.
}
