% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{marginal.mle_fit}
\alias{marginal.mle_fit}
\title{Method for obtaining the marginal distribution of an MLE
that is based on asymptotic assumptions:}
\usage{
\method{marginal}{mle_fit}(x, indices)
}
\arguments{
\item{x}{The distribution object.}

\item{indices}{The indices of the marginal distribution to obtain.}
}
\value{
An \code{mle_fit} object representing the marginal distribution for the
  selected parameter indices.
}
\description{
`x ~ MVN(params(x), inv(H)(x))`
}
\details{
where H is the (observed or expecation) Fisher information
matrix.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), loglike = -120, nobs = 100L)
marginal(fit, 1)
}
