% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{mean.mle_boot}
\alias{mean.mle_boot}
\title{Mean of bootstrap replicates.}
\usage{
\method{mean}{mle_boot}(x, ...)
}
\arguments{
\item{x}{An \code{mle_boot} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Numeric vector of column means of bootstrap replicates.
}
\description{
Returns the empirical mean of the bootstrap replicates (which includes
bootstrap bias). For the bias-corrected point estimate, use \code{params()}.
}
