% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mean.mle_fit}
\alias{mean.mle_fit}
\title{Mean of the asymptotic distribution of an MLE.}
\usage{
\method{mean}{mle_fit}(x, ...)
}
\arguments{
\item{x}{An \code{mle_fit} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Numeric vector of parameter estimates.
}
\description{
Returns the parameter estimates, which are the mean of the asymptotic
distribution.
}
