% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_rmap.R
\name{mle_apply}
\alias{mle_apply}
\title{\code{rmap.mle} computes the distribution of \code{f(x)} where \code{x}
is an \code{mle} object.}
\usage{
mle_apply(x, g, method = "delta", keep_obs = F)
}
\arguments{
\item{x}{an \code{mle} object.}

\item{g}{a function that accepts objects like \code{point(x)}.}
}
\description{
By the invariance property of the MLE, if \code{x} is an \code{mle} object,
then under the right conditions, asymptotically, \code{f(x)} is normally
distributed with an approximation given by
\code{f(x) ~ normal(f(point(x)),sigma)} where \code{sigma} is
the variance-covariance of \code{f(x)}, e.g., the sample
covariance of \code{f(x1),...f(xn)} where \code{xj} is sampled from
\code{sampler(x)}.
}
\details{
delta method
bootstrap method
monte carlo method
}
