% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{mle_boot}
\alias{mle_boot}
\title{A function for estimating the empirical sampling distribution the
MLE using the Bootstrap method.}
\usage{
mle_boot(mle, mle_solver, data = NULL, R = NULL, ...)
}
\arguments{
\item{mle}{a fitted \code{mle} object.}

\item{mle_solver}{given \code{data} find the MLE}

\item{data}{data for generatinng MLEs for the bootstrap resampling; default
is NULL. If NULL, then try to use \code{obs(x)} instead.}

\item{R}{bootstrap replicates}

\item{...}{additional arguments to pass.}
}
\description{
A function for estimating the empirical sampling distribution the
MLE using the Bootstrap method.
}
