% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_confint}
\alias{mle_confint}
\title{Function to compute the confidence intervals of the parameters of
\code{mle} objects.}
\usage{
mle_confint(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the \code{mle} object to compute the confidence intervals for.}

\item{parm}{?}

\item{level}{confidence level, defaults to 0.95 (alpha=.05).}

\item{...}{additional arguments to pass.}
}
\description{
Function to compute the confidence intervals of the parameters of
\code{mle} objects.
}
