% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{mle_exp}
\alias{mle_exp}
\title{Maximum Likelihood Estimation (MLE) for the Exponential Distribution}
\usage{
mle_exp(x, keep_obs = FALSE)
}
\arguments{
\item{x}{a numeric vector representing a sample of observations.}

\item{keep_obs}{logical. If TRUE, the observations are stored with the `mle`
object. Default is FALSE.}
}
\value{
Returns an `mle` object that contains:
}
\description{
This function computes the Maximum Likelihood Estimation (MLE) of the rate
parameter for a given sample that is assumed to be independently and
identically distributed (i.i.d) and drawn from the Exponential distribution.
}
\examples{
\dontrun{
  obs <- rexp(100, rate = 0.5)
  rate.hat <- mle_exp(obs)
  summary(rate.hat)
}

}
