% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_fisher_info}
\alias{mle_fisher_info}
\title{Function for obtaining the fisher information matrix of an \code{mle} object.}
\usage{
mle_fisher_info(x, ...)
}
\arguments{
\item{x}{the \code{mle} object to obtain the fisher information of.}

\item{...}{unused by \code{mle} objects. particular specializations of
\code{mle} objects may use it, however.}
}
\description{
Function for obtaining the fisher information matrix of an \code{mle} object.
}
