% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_mle.R
\name{mle_gradient_ascent}
\alias{mle_gradient_ascent}
\title{mle_gradient_ascent}
\usage{
mle_gradient_ascent(ll, theta0, score, options)
}
\arguments{
\item{ll}{log-likelihood function}

\item{theta0}{initial guess of theta with `p` components}

\item{score}{score function of type `R^p -> R^p`, the gradient of `ll`}
}
\description{
MLE method using gradient ascent
}
