% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_numerical.R
\name{mle_iterative}
\alias{mle_iterative}
\title{General iterative MLE method}
\usage{
mle_iterative(
  l,
  theta0,
  dir = NULL,
  stop_cond = NULL,
  sup = NULL,
  eta = 1,
  r = 0.5,
  max_iter = 0L,
  debug = F
)
}
\arguments{
\item{l}{log-likelihood function of type \eqn{R^p \mapsto R}}

\item{theta0}{initial guess of \eqn{\theta} with \eqn{p} components}

\item{dir}{promising direction function of type \eqn{R^p \mapsto R^p}}

\item{stop_cond}{stopping condition function of type \eqn{R^p \times R^p \mapsto \{T,F\}}}

\item{sup}{domain of support function of type \eqn{R^p \mapsto \{T,F\}} for the log-likelihood function \eqn{l}}

\item{eta}{learning rate, defaults to 1}

\item{r}{backing tracking parameter}

\item{max_iter}{maximum number of iterations}

\item{debug}{Boolean, output debugging information if TRUE; defaults to FALSE}
}
\description{
General iterative MLE method
}
