% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_linear_transform.R
\name{mle_linear_transform}
\alias{mle_linear_transform}
\title{Linearly transform an \code{mle} object \code{x} by multiplying it
on the LHS by a matrix \code{A}.}
\usage{
mle_linear_transform(A, x)
}
\arguments{
\item{A}{a (non-random) matrix}

\item{x}{an \code{mle} object to linearly transform}
}
\description{
Linearly transform an \code{mle} object \code{x} by multiplying it
on the LHS by a matrix \code{A}.
}
