% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_mle.R
\name{mle_optim}
\alias{mle_optim}
\title{mle_optim}
\usage{
mle_optim(sol)
}
\arguments{
\item{sol}{the output of `optim`}
}
\value{
a `numerical_mle` object, specialized for `optim` (stats package)
solutions.
}
\description{
This function takes the output of `optim` and turns it into an `mle` object.
}
