% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_point}
\alias{mle_point}
\title{Computes the point estimate of an mle object.}
\usage{
mle_point(x, ...)
}
\arguments{
\item{x}{the mle object(s).}

\item{...}{unused by \code{mle} objects. particular specializations of
\code{mle} objects may use it, however.}
}
\description{
Computes the point estimate of an mle object.
}
