% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{mle_samp_bs_loglike}
\alias{mle_samp_bs_loglike}
\title{A function for computing the sampling distribution of a statistic of the
MLE's sampling distribution using the Bootstrap method.}
\usage{
mle_samp_bs_loglike(x, loglike, data = NULL, R = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{mle} object.}

\item{loglike}{a generator for the log-likelihood function; it accepts
observations and constructs the log-likelihood function
with the \code{data} fixed and as a function of theta.}

\item{data}{data for generatinng MLEs for the bootstrap resampling; default
is NULL. If NULL, then try to use \code{obs(x)} instead.}

\item{R}{bootstrap replicates}

\item{...}{additional arguments to pass.}
}
\description{
A function for computing the sampling distribution of a statistic of the
MLE's sampling distribution using the Bootstrap method.
}
