% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_sampler}
\alias{mle_sampler}
\title{Method for sampling from an \code{mle} object.}
\usage{
mle_sampler(x, ...)
}
\arguments{
\item{x}{the \code{mle} object to create sampler for}

\item{...}{additional arguments to pass to \code{mle} objects, like sampling
method, \code{?mvtnorm}.}
}
\description{
It creates a sampler for the \code{mle} object. It returns a function
that accepts a single parameter \code{n} denoting the number of samples
to draw from the \code{mle} object.
}
