% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{mle_sols}
\alias{mle_sols}
\title{Generic method for obtaining the best solutions for an estimator, e.g.,
maximum likelihood estimation may have many possible solutions that satisfy
the maximum likelihood equation, such that
\code{argmax(x) = {L(x) : x in solution set}}
is a set: empty set, singleton set (unique), or a set with more than one
element.}
\usage{
mle_sols(x, ...)
}
\arguments{
\item{x}{the MLE object to obtain the solutions for}

\item{...}{additional arguments to pass}
}
\value{
a set of points satisfying the MLE equations
#' @export
}
\description{
Generic method for obtaining the best solutions for an estimator, e.g.,
maximum likelihood estimation may have many possible solutions that satisfy
the maximum likelihood equation, such that
\code{argmax(x) = {L(x) : x in solution set}}
is a set: empty set, singleton set (unique), or a set with more than one
element.
}
