% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_numerical.R
\name{mle_solver_random_restarts}
\alias{mle_solver_random_restarts}
\title{Iterative MLE method using gradient ascent}
\usage{
mle_solver_random_restarts(mle_solver, start_gen, trials = 10)
}
\arguments{
\item{mle_solver}{MLE solver, takes a starting point}

\item{start_gen}{starting point generator}

\item{trials}{number of trials to run}
}
\description{
Iterative MLE method using gradient ascent
}
