% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_sum}
\alias{mle_sum}
\title{Compute the inverse-variance weighted MLE of the sum of MLE objects.}
\usage{
mle_sum(..., na.rm = FALSE)
}
\arguments{
\item{...}{the list of \code{mle} objects to sum over.}

\item{na.rm}{unused when summing over \code{mle} objects.}
}
\description{
Compute the inverse-variance weighted MLE of the sum of MLE objects.
}
