% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weibull.R
\name{mle_weibull}
\alias{mle_weibull}
\title{MLE of the rate parameter when we assume the sample is i.i.d. and drawn
from the weibull distribution.}
\usage{
mle_weibull(x, k0 = 1, rel_tol = 1e-07, keep_obs = F)
}
\arguments{
\item{x}{a sample of observations}

\item{rel_tol}{when numerically solving the MLE equation,
`|new-old|/|new| <= rel_tol` is stopping condition}

\item{keep_obs}{Boolean, specifies whether to keep observations}

\item{k}{initial estimate of shape parameter k}
}
\value{
an `mle` object.
}
\description{
MLE of the rate parameter when we assume the sample is i.i.d. and drawn
from the weibull distribution.
}
