% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weighted.R
\name{mle_weighted_2}
\alias{mle_weighted_2}
\title{Accepts a list of \code{mle} objects for some parameter, say \code{theta},
and combines them into a single estimator \code{mle_weighted}.}
\usage{
mle_weighted_2(mles)
}
\arguments{
\item{mles}{A list of \code{mle} objects, all for the same parameter.}
}
\value{
an object of type \code{mle_weighted} which inherits from \code{mle}.
}
\description{
It combines the \code{mle} objects by adding them together, weighted by
the inverse of their respective variance-covariance matrix (information
matrix). Intuitively, the higher the variance, the less weight an \code{mle}
is given in the summation, or alternatively, the more information it
has about the parameter, the more weight it is given in the summation.
}
\details{
Each \code{mle} object should have a \code{fim} method, which returns
the information matrix (inverse of the variance-covariance matrix).
}
