% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mse.mle}
\alias{mse.mle}
\title{Computes the MSE of an MLE}
\usage{
\method{mse}{mle}(x, ...)
}
\arguments{
\item{x}{the \code{mle} object to compute the MSE of.}

\item{...}{this is not used for object type \code{mle}.}
}
\description{
The MSE of an estimator \code{theta.hat} is defined as
    \code{mse(theta.hat) = E[(theta.hat-theta)^2] = vcov(theta.hat) + (bias(theta.hat))^2}.
where bias = E(theta.hat-theta), i.e., the MSE depends on theta.
The definition of the MSE of \code{theta.hat} is given by
the difference between the
$$
    f(theta.hat)
$$
}
