% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{mse.mle_fit_boot}
\alias{mse.mle_fit_boot}
\title{Computes the estimate of the MSE of a `boot` object.}
\usage{
\method{mse}{mle_fit_boot}(x, theta = NULL, ...)
}
\arguments{
\item{x}{the `boot` object to compute the MSE of.}

\item{theta}{true parameter value (not used for `mle_boot`)}

\item{...}{pass additional arguments into `vcov`}
}
\value{
The MSE matrix estimated from bootstrap variance and bias.
}
\description{
Computes the estimate of the MSE of a `boot` object.
}
\examples{
set.seed(1)
b <- boot::boot(rexp(50, 2), function(d, i) 1/mean(d[i]), R = 99)
fit <- mle_boot(b)
mse(fit)
}
