% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{nobs.boot}
\alias{nobs.boot}
\title{Method for obtaining the number of observations in the sample used by
an \code{mle}.}
\usage{
\method{nobs}{boot}(object, ...)
}
\arguments{
\item{object}{the \code{mle} object to obtain the number of observations for}

\item{...}{additional arguments to pass}
}
\description{
Method for obtaining the number of observations in the sample used by
an \code{mle}.
}
