% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{nobs}
\alias{nobs}
\title{Generic method for obtaining the number of observations used by an MLE fit.}
\usage{
nobs(x)
}
\arguments{
\item{x}{the object to compute the number of observations for}
}
\description{
Generic method for obtaining the number of observations used by an MLE fit.
}
