% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{normal_fim}
\alias{normal_fim}
\title{Fisher information matrix generator given data `x` for the normal
distribution}
\usage{
normal_fim(x, observed = TRUE)
}
\arguments{
\item{x}{data (simple random sample)}

\item{observed}{whether to return the observed fisher information, default is
`TRUE`}
}
\description{
Fisher information matrix generator given data `x` for the normal
distribution
}
