% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal_mu_var.R
\name{normal_fisher_info_mu_var}
\alias{normal_fisher_info_mu_var}
\title{Fisher information matrix generator given data \code{x} for the normal
distribution}
\usage{
normal_fisher_info_mu_var(x, observed = T)
}
\arguments{
\item{x}{data (simple random sample)}

\item{observed}{whether to return the observed fisher information, default is TRUE}
}
\description{
Fisher information matrix generator given data \code{x} for the normal
distribution
}
