% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{normal_loglike}
\alias{normal_loglike}
\title{log-likelihood function generator given data `x` for the normal
distribution}
\usage{
normal_loglike(x)
}
\arguments{
\item{x}{data}
}
\description{
log-likelihood function generator given data `x` for the normal
distribution
}
