% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal_mu_var.R
\name{normal_loglike_mu_var}
\alias{normal_loglike_mu_var}
\title{log-likelihood function generator given data \code{x} for the normal
distribution}
\usage{
normal_loglike_mu_var(x)
}
\arguments{
\item{x}{data}
}
\description{
log-likelihood function generator given data \code{x} for the normal
distribution
}
