% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{normal_score}
\alias{normal_score}
\title{Score function generator given data \code{x} for the normal
distribution}
\usage{
normal_score(x)
}
\arguments{
\item{x}{data (simple random sample)}
}
\description{
Score function generator given data \code{x} for the normal
distribution
}
