% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal_mu_var.R
\name{normal_score_mu_var}
\alias{normal_score_mu_var}
\title{Score function generator given data \code{x} for the normal
distribution}
\usage{
normal_score_mu_var(x)
}
\arguments{
\item{x}{data (simple random sample)}
}
\description{
Score function generator given data \code{x} for the normal
distribution
}
