% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{nparams.boot}
\alias{nparams.boot}
\title{Method for obtaining the number of parameters of an \code{boot} object.}
\usage{
\method{nparams}{boot}(x)
}
\arguments{
\item{x}{the \code{boot} object to obtain the number of parameters of}
}
\description{
Method for obtaining the number of parameters of an \code{boot} object.
}
