% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{nparams.mle_boot}
\alias{nparams.mle_boot}
\title{Method for obtaining the number of parameters of an `boot` object.}
\usage{
\method{nparams}{mle_boot}(x)
}
\arguments{
\item{x}{the `boot` object to obtain the number of parameters of}
}
\description{
Method for obtaining the number of parameters of an `boot` object.
}
