% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_emp.R
\name{nparams.mle_emp}
\alias{nparams.mle_emp}
\title{Method for obtaining the number of parameters of an `mle_emp` object.}
\usage{
\method{nparams}{mle_emp}(x)
}
\arguments{
\item{x}{the `mle_emp` object to obtain the number of parameters of}
}
\description{
Method for obtaining the number of parameters of an `mle_emp` object.
}
