% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{nparams}
\alias{nparams}
\title{Generic method for obtaining the number of parameters of a fitted
distribution object.}
\usage{
nparams(x)
}
\arguments{
\item{x}{the fitted object to obtain the number of parameters for}
}
\description{
Generic method for obtaining the number of parameters of a fitted
distribution object.
}
