% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{obs.boot}
\alias{obs.boot}
\title{Method for obtaining the observations used by the \code{mle}.}
\usage{
\method{obs}{boot}(object, ...)
}
\arguments{
\item{object}{the \code{mle} object to obtain the number of observations for}

\item{...}{additional arguments to pass}
}
\description{
Method for obtaining the observations used by the \code{mle}.
}
