% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{obs.mle_fit}
\alias{obs.mle_fit}
\title{Method for obtaining the observations used by the `mle_fit` object `x`.}
\usage{
\method{obs}{mle_fit}(x)
}
\arguments{
\item{x}{the `mle_fit` object to obtain the number of observations for}
}
\value{
The observation data used to fit the MLE, or NULL if not stored.
}
\description{
Method for obtaining the observations used by the `mle_fit` object `x`.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), obs = rnorm(100), nobs = 100L)
head(obs(fit))
}
