% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{obs.mle_fit_boot}
\alias{obs.mle_fit_boot}
\title{Method for obtaining the observations used by the `mle_fit_boot`.}
\usage{
\method{obs}{mle_fit_boot}(x)
}
\arguments{
\item{x}{the `mle_fit_boot` object to obtain the number of observations for}
}
\value{
The original data used for bootstrapping.
}
\description{
Method for obtaining the observations used by the `mle_fit_boot`.
}
\examples{
set.seed(1)
b <- boot::boot(rexp(50, 2), function(d, i) 1/mean(d[i]), R = 99)
fit <- mle_boot(b)
head(obs(fit))
}
