% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{observed_fim.mle}
\alias{observed_fim.mle}
\title{Function for obtaining the observed FIM of an `mle` object.}
\usage{
\method{observed_fim}{mle}(x, ...)
}
\arguments{
\item{x}{the `mle` object to obtain the FIM of.}

\item{...}{pass additional arguments}
}
\description{
Function for obtaining the observed FIM of an `mle` object.
}
