% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{observed_fim.mle_fit}
\alias{observed_fim.mle_fit}
\title{Function for obtaining the observed FIM of an `mle_fit` object.}
\usage{
\method{observed_fim}{mle_fit}(x, ...)
}
\arguments{
\item{x}{the `mle_fit` object to obtain the FIM of.}

\item{...}{pass additional arguments}
}
\value{
The observed Fisher Information Matrix, or NULL if not available.
}
\description{
Function for obtaining the observed FIM of an `mle_fit` object.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), info = solve(diag(c(0.04, 0.32))),
  nobs = 100L)
observed_fim(fit)
}
