% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.mle.R
\name{orthogonal}
\alias{orthogonal}
\title{Generic method for determining the orthogonal parameters of an estimator.}
\usage{
orthogonal(x, tol, ...)
}
\arguments{
\item{x}{the estimator}

\item{tol}{the tolerance for determining if a number is close enough to zero}

\item{...}{additional arguments to pass}
}
\value{
Logical vector or matrix indicating which parameters are orthogonal.
}
\description{
Generic method for determining the orthogonal parameters of an estimator.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), info = solve(diag(c(0.04, 0.32))),
  loglike = -120, nobs = 100L)
orthogonal(fit)
}
