% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{params.boot}
\alias{params.boot}
\title{Method for obtaining the parameters of an \code{boot} object.}
\usage{
\method{params}{boot}(x)
}
\arguments{
\item{x}{the \code{boot} object to obtain the parameters of.}
}
\description{
Method for obtaining the parameters of an \code{boot} object.
}
