% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/mle.R
\name{params.mle}
\alias{params.mle}
\title{Method for obtaining the parameters of a fitted object.}
\usage{
params.mle(x)

params.mle(x)
}
\arguments{
\item{x}{the \code{mle} object to obtain the parameters of}
}
\description{
Method for obtaining the parameters of a fitted object.

Method for obtaining the parameters of an \code{mle} object.
}
