% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_emp.R
\name{params.mle_emp}
\alias{params.mle_emp}
\title{Method for obtaining the parameters of an `mle_emp` object.}
\usage{
\method{params}{mle_emp}(x)
}
\arguments{
\item{x}{the `mle_emp` object to obtain the parameters of.}
}
\description{
Method for obtaining the parameters of an `mle_emp` object.
}
