% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{params.mle_fit}
\alias{params.mle_fit}
\title{Method for obtaining the parameters of an `mle_fit` object.}
\usage{
\method{params}{mle_fit}(x)
}
\arguments{
\item{x}{the `mle_fit` object to obtain the parameters of}
}
\value{
Numeric vector of parameter estimates.
}
\description{
Method for obtaining the parameters of an `mle_fit` object.
}
