% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algeabraic.mle.R
\name{params}
\alias{params}
\title{Generic method for obtaining the parameters of a fitted MLE object.}
\usage{
params(x, ...)
}
\arguments{
\item{x}{the object to obtain the parameters of}

\item{...}{additional arguments to pass}
}
\description{
Generic method for obtaining the parameters of a fitted MLE object.
}
