% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{point.exp_mle}
\alias{point.exp_mle}
\title{Computes the point estimate of an mle object.}
\usage{
\method{point}{exp_mle}(x, ...)
}
\arguments{
\item{x}{the mle object(s).}

\item{...}{unused by \code{mle} objects. particular specializations of
\code{mle} objects may use it, however.}
}
\description{
Computes the point estimate of an mle object.
}
