% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{point.mle}
\alias{point.mle}
\title{Computes the point estimate of an `mle` object.}
\usage{
\method{point}{mle}(x, ...)
}
\arguments{
\item{x}{the `mle` object.}

\item{...}{pass additional arguments}
}
\description{
Computes the point estimate of an `mle` object.
}
