% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{point.mle_boot}
\alias{point.mle_boot}
\title{Computes the point estimate of an \code{mle} object.}
\usage{
\method{point}{mle_boot}(x, ...)
}
\arguments{
\item{x}{the \code{boot} object.}

\item{...}{pass additional arguments}
}
\description{
Computes the point estimate of an \code{mle} object.
}
