% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_emp.R
\name{point.mle_emp}
\alias{point.mle_emp}
\title{Computes the point estimate of an `mle` object.}
\usage{
\method{point}{mle_emp}(x, ...)
}
\arguments{
\item{x}{the `mle` object to compute the point estimate of}

\item{...}{pass additional arguments}
}
\description{
Computes the point estimate of an `mle` object.
}
