% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{point}
\alias{point}
\title{Generic method for obtaining the best point estimate from an estimator
object `x`.}
\usage{
point(x, ...)
}
\arguments{
\item{x}{the object to obtain the point estimate of}

\item{...}{additional arguments to pass}
}
\description{
Generic method for obtaining the best point estimate from an estimator
object `x`.
}
