% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.interval}
\alias{pred.interval}
\title{Function for obtaining an empirical sampling distribution from the asymptotic
distribution of the \code{mle} object \code{x} when applied to a function
\code{g}.}
\usage{
pred.interval(x, g, n = 1000, alpha = 0.05, ...)
}
\arguments{
\item{x}{the \code{mle} object}

\item{g}{the function to predict the value of with respect to input \code{x}}

\item{n}{the sample size}
}
\description{
Function for obtaining an empirical sampling distribution from the asymptotic
distribution of the \code{mle} object \code{x} when applied to a function
\code{g}.
}
