% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.mle.R
\name{pred}
\alias{pred}
\title{Generic method for computing the predictive confidence interval given an estimator object `x`.}
\usage{
pred(x, samp = NULL, alpha = 0.05, ...)
}
\arguments{
\item{x}{the estimator object}

\item{samp}{a sampler for random variable that is parameterized by mle `x`}

\item{alpha}{(1-alpha)/2 confidence interval}

\item{...}{additional arguments to pass}
}
\description{
Generic method for computing the predictive confidence interval given an estimator object `x`.
}
