% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{print.mle}
\alias{print.mle}
\title{Print method for `mle` objects.}
\usage{
\method{print}{mle}(x, ...)
}
\arguments{
\item{x}{the `mle` object to print}

\item{...}{additional arguments to pass}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for `mle` objects.
}
