% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{print.summary_mle}
\alias{print.summary_mle}
\title{Function for printing a \code{summary} object for an \code{mle} object.}
\usage{
\method{print}{summary_mle}(object, ...)
}
\arguments{
\item{object}{the \code{summary_mle} object}

\item{...}{pass additional arguments}
}
\description{
Function for printing a \code{summary} object for an \code{mle} object.
}
