% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{resid.mle}
\alias{resid.mle}
\title{A function for computing the residuals of \code{object}, which is
a fitted \code{mle} object.}
\usage{
\method{resid}{mle}(object, ...)
}
\arguments{
\item{object}{a fitted \code{mle} object.}

\item{...}{additional arguments to pass.}
}
\description{
A function for computing the residuals of \code{object}, which is
a fitted \code{mle} object.
}
