% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{rmap}
\alias{rmap}
\title{Generic function for computing the distribution of \code{f(x)} where \code{x}
models a distribution (random element) object.}
\usage{
rmap(x, g, n, ...)
}
\arguments{
\item{x}{a list of distribution (random element) objects.}

\item{g}{a function that accepts arguments sampled from the distribution
objects in \code{x}.}

\item{n}{number of samples for the distribution to draw from to estimate \code{f(x)}}

\item{...}{additional arguments to pass.}
}
\description{
When we apply a function \code{f} to a distribution \code{x}, the result
is a distribution object \code{f(x)}.
}
