% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{sample_size.mle}
\alias{sample_size.mle}
\title{Method for determining the sample size of an \code{mle} object.}
\usage{
\method{sample_size}{mle}(x)
}
\arguments{
\item{x}{the \code{mle} object}
}
\description{
Method for determining the sample size of an \code{mle} object.
}
