% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{sample_size}
\alias{sample_size}
\title{Method for determining the sample size of an estimator object.}
\usage{
sample_size(x)
}
\arguments{
\item{x}{the estimator}
}
\description{
Method for determining the sample size of an estimator object.
}
