% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{sampler.boot}
\alias{sampler.boot}
\title{Method for sampling from an \code{boot} object.}
\usage{
\method{sampler}{boot}(x, ...)
}
\arguments{
\item{x}{the \code{boot} object to create sampler for}

\item{...}{additional arguments to pass}
}
\description{
It creates a sampler for the \code{boot} object. It returns a function
that accepts a single parameter \code{n} denoting the number of samples
to draw from the \code{boot} object.
}
