% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{sampler.mle}
\alias{sampler.mle}
\title{Method for sampling from an \code{mle} object.}
\usage{
\method{sampler}{mle}(x, p = 1, ...)
}
\arguments{
\item{x}{the \code{mle} object to create sampler for}

\item{p}{the value that defines the \code{p}-confidence region sampled from, defaults to \code{1}}

\item{...}{additional arguments to pass}
}
\description{
It creates a sampler for the \code{mle} object. It returns a function
that accepts a single parameter \code{n} denoting the number of samples
to draw from the \code{mle} object.
}
