% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_emp.R
\name{sampler.mle_emp}
\alias{sampler.mle_emp}
\title{Method for sampling from an `mle_emp` object.}
\usage{
\method{sampler}{mle_emp}(x, ...)
}
\arguments{
\item{x}{the `mle_emp` object to create sampler for}

\item{...}{additional arguments to pass}
}
\description{
Method for sampling from an `mle_emp` object.
}
