% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{sampler.mle_exp}
\alias{sampler.mle_exp}
\title{Method for sampling from an \code{mle_exp} object.}
\usage{
\method{sampler}{mle_exp}(x, ...)
}
\arguments{
\item{x}{the \code{mle_exp} object to create sampler for.}

\item{...}{additional arguments to pass to \code{mle_exp} objects.}
}
\description{
It creates a sampler for the \code{mle_exp} object. It returns a function
that accepts a single parameter \code{n} denoting the number of samples
to draw from the \code{mle_exp} object.
}
