% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{sampler.mle_fit_boot}
\alias{sampler.mle_fit_boot}
\title{Method for sampling from an `mle_fit_boot` object.}
\usage{
\method{sampler}{mle_fit_boot}(x, ...)
}
\arguments{
\item{x}{the `mle_fit_boot` object to create sampler for}

\item{...}{additional arguments to pass (not used)}
}
\value{
A function that takes parameter \code{n} and returns \code{n} samples
  drawn from the bootstrap replicates.
}
\description{
It creates a sampler for the `mle_fit_boot` object. It returns a function
that accepts a single parameter `n` denoting the number of samples
to draw from the `mle_fit_boot` object.
}
\details{
Unlike the `sampler` method for the more general `mle_fit` objects,
for `mle_fit_boot` objects, we sample from the bootstrap replicates, which
are more representative of the sampling distribution, particularly for
small samples.
}
\examples{
set.seed(1)
b <- boot::boot(rexp(50, 2), function(d, i) 1/mean(d[i]), R = 99)
fit <- mle_boot(b)
sampler(fit)(5)
}
