% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{sampler}
\alias{sampler}
\title{Generic function for sampling from distribution objects.}
\usage{
sampler(x, ...)
}
\arguments{
\item{x}{the `x` object to create a sampler for}

\item{...}{additional arguments to pass}
}
\description{
It creates a sampler for the `x` object. It returns a function
that accepts a single parameter `n` denoting the number of samples
to draw from the `x` object.
}
