% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{score.mle}
\alias{score.mle}
\title{Computes the score of an \code{mle} object.}
\usage{
\method{score}{mle}(x, ...)
}
\arguments{
\item{x}{the \code{mle} object to compute the score of.}

\item{...}{additional arguments to pass}
}
\description{
If reguarlity conditions are satisfied, it should be zero (or approximately,
if rounding errors occur).
}
