% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{score}
\alias{score}
\title{score}
\usage{
score(x, ...)
}
\arguments{
\item{x}{the object to compute the score of.}

\item{...}{pass additional arguments}
}
\description{
Generic function for computing the score of an estimator
object with respect to its log-likelihood.
}
