% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{score_val.mle}
\alias{score_val.mle}
\title{Computes the score of an `mle` object (score evaluated at the MLE).}
\usage{
\method{score_val}{mle}(x, ...)
}
\arguments{
\item{x}{the `mle` object to compute the score of.}

\item{...}{additional arguments to pass (not used)}
}
\description{
If reguarlity conditions are satisfied, it should be zero (or approximately,
if rounding errors occur).
}
