% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{score_val.mle_fit}
\alias{score_val.mle_fit}
\title{Computes the score of an `mle_fit` object (score evaluated at the MLE).}
\usage{
\method{score_val}{mle_fit}(x, ...)
}
\arguments{
\item{x}{the `mle_fit` object to compute the score of.}

\item{...}{additional arguments to pass (not used)}
}
\value{
The score vector evaluated at the MLE, or NULL if not available.
}
\description{
If reguarlity conditions are satisfied, it should be zero (or approximately,
if rounding errors occur).
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), score = c(0.001, -0.002),
  nobs = 100L)
score_val(fit)
}
