% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.mle.R
\name{score_val}
\alias{score_val}
\title{Generic method for computing the score of an estimator
object (gradient of its log-likelihood function evaluated
at the MLE).}
\usage{
score_val(x, ...)
}
\arguments{
\item{x}{the object to compute the score of.}

\item{...}{pass additional arguments}
}
\value{
The score vector evaluated at the MLE.
}
\description{
Generic method for computing the score of an estimator
object (gradient of its log-likelihood function evaluated
at the MLE).
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), score = c(0.001, -0.002),
  loglike = -120, nobs = 100L)
score_val(fit)
}
