% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{sd.mle}
\alias{sd.mle}
\title{Method for obtaining the standard deviation of an \code{mle} object.}
\usage{
sd.mle(x, na.rm = FALSE)
}
\arguments{
\item{x}{the \code{mle} object}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Method for obtaining the standard deviation of an \code{mle} object.
}
