% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{se.mle}
\alias{se.mle}
\title{Function for obtaining an estimate of the standard error of the MLE
object `x`.}
\usage{
\method{se}{mle}(x, se.matrix = FALSE, ...)
}
\arguments{
\item{x}{the MLE object}

\item{se.matrix}{if `TRUE`, return the square root of the variance-covariance}

\item{...}{additional arguments to pass (not used)}
}
\value{
Vector of standard errors, or matrix if \code{se.matrix = TRUE}, or
  NULL if variance-covariance is not available.
}
\description{
Function for obtaining an estimate of the standard error of the MLE
object `x`.
}
