% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{se.mle_boot}
\alias{se.mle_boot}
\title{Function for obtaining an estimate of the standard error of the bootstrap of
the MLE \code{object}.}
\usage{
\method{se}{mle_boot}(object)
}
\arguments{
\item{object}{the bootstrapped MLE object}
}
\description{
Function for obtaining an estimate of the standard error of the bootstrap of
the MLE \code{object}.
}
