% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{se.mle_fit}
\alias{se.mle_fit}
\title{Function for obtaining an estimate of the standard error of the MLE
object `x`.}
\usage{
\method{se}{mle_fit}(x, se.matrix = FALSE, ...)
}
\arguments{
\item{x}{the MLE object}

\item{se.matrix}{if `TRUE`, return the square root of the variance-covariance}

\item{...}{additional arguments to pass (not used)}
}
\value{
Vector of standard errors, or matrix if \code{se.matrix = TRUE}, or
  NULL if variance-covariance is not available.
}
\description{
Function for obtaining an estimate of the standard error of the MLE
object `x`.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), loglike = -120, nobs = 100L)
se(fit)
}
