% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{sum.mle}
\alias{sum.mle}
\title{Compute the inverse-variance weighted MLE of the sum of MLE objects.}
\usage{
\method{sum}{mle}(..., na.rm = FALSE)
}
\arguments{
\item{...}{the list of \code{mle} objects to sum over.}

\item{na.rm}{unused when summing over \code{mle} objects.}
}
\value{
an object of type \code{mle_weighted}.
}
\description{
Compute the inverse-variance weighted MLE of the sum of MLE objects.
}
