% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{summary.boot}
\alias{summary.boot}
\title{Function for obtaining a summary of \code{object}, which is a fitted
\code{boot} object.}
\usage{
\method{summary}{boot}(object, ...)
}
\arguments{
\item{object}{the \code{boot} object}

\item{...}{pass additional arguments}
}
\description{
Function for obtaining a summary of \code{object}, which is a fitted
\code{boot} object.
}
