% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{summary.mle}
\alias{summary.mle}
\title{Function for obtaining a summary of an \code{mle} object.}
\usage{
\method{summary}{mle}(object, ...)
}
\arguments{
\item{object}{the \code{mle} object}

\item{...}{pass additional arguments}
}
\description{
Function for obtaining a summary of an \code{mle} object.
}
