% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{summary.mle_fit}
\alias{summary.mle_fit}
\title{Function for obtaining a summary of `object`, which is a fitted
`mle_fit` object.}
\usage{
\method{summary}{mle_fit}(object, ...)
}
\arguments{
\item{object}{the `mle_fit` object}

\item{...}{pass additional arguments}
}
\value{
An object of class \code{summary_mle_fit}.
}
\description{
Function for obtaining a summary of `object`, which is a fitted
`mle_fit` object.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), loglike = -120, nobs = 100L)
summary(fit)
}
