% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{sup.mle}
\alias{sup.mle}
\title{Support of the asymptotic distribution of an MLE.}
\usage{
\method{sup}{mle}(x)
}
\arguments{
\item{x}{An \code{mle} object.}
}
\value{
A support object (interval for normal distributions).
}
\description{
Support of the asymptotic distribution of an MLE.
}
