% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{two.sample.test}
\alias{two.sample.test}
\title{General two sample hypothesis test.}
\usage{
two.sample.test(sampler, obs, cdf = NULL, options = list(), ...)
}
\arguments{
\item{sampler}{random sampler for reference distribution}

\item{obs}{a sample of observations}

\item{cdf}{a cdf for the reference distribution, defaults to NULL}

\item{n}{sample size (for sampler)}

\item{method}{the method for the two sample hypothesis test}
}
\value{
hypothesis test object
}
\description{
General two sample hypothesis test.
}
