% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal_mu_var.R
\name{unbias.mle_normal_mu_var}
\alias{unbias.mle_normal_mu_var}
\title{Compute the unbiased point estimate from a \code{mle_normal}
estimator object.}
\usage{
\method{unbias}{mle_normal_mu_var}(x, ...)
}
\arguments{
\item{x}{the \code{mle_normal} from which to compute an unbiased estimate}

\item{...}{unused additional arguments}
}
\description{
Compute the unbiased point estimate from a \code{mle_normal}
estimator object.
}
