% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{unbias}
\alias{unbias}
\title{Compute the unbiased point estimate given an estimator object \code{x}.}
\usage{
unbias(x, ...)
}
\arguments{
\item{x}{the object from which to compute an unbiased estimate}

\item{...}{additional arguments to pass}
}
\description{
Compute the unbiased point estimate given an estimator object \code{x}.
}
