% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{unbiased_point}
\alias{unbiased_point}
\title{Function for retrieving an unbiased point estimate from an MLE object}
\usage{
unbiased_point(x, ...)
}
\arguments{
\item{x}{the \code{mle} object}

\item{...}{additional arguments to pass}
}
\description{
Function for retrieving an unbiased point estimate from an MLE object
}
