% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{vcov.boot}
\alias{vcov.boot}
\title{Computes the variance-covariance matrix of \code{boot} object.}
\usage{
\method{vcov}{boot}(object, ...)
}
\arguments{
\item{object}{the \code{boot} object to obtain the variance-covariance of}

\item{...}{additional arguments to pass}
}
\description{
Computes the variance-covariance matrix of \code{boot} object.
}
