% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{vcov.mle}
\alias{vcov.mle}
\title{Computes the variance-covariance matrix of `mle` object `x`.}
\usage{
\method{vcov}{mle}(x, ...)
}
\arguments{
\item{x}{the `mle` object to obtain the variance-covariance of}

\item{...}{additional arguments to pass}
}
\description{
Computes the variance-covariance matrix of `mle` object `x`.
}
