% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_emp.R
\name{vcov.mle_emp}
\alias{vcov.mle_emp}
\title{Computes the variance-covariance matrix of `mle_emp` object.}
\usage{
\method{vcov}{mle_emp}(x, ...)
}
\arguments{
\item{x}{the `mle_emp` object to obtain the variance-covariance of}

\item{...}{additional arguments to pass}
}
\description{
Computes the variance-covariance matrix of `mle_emp` object.
}
