% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weibull.R
\name{weibull_fisher_info}
\alias{weibull_fisher_info}
\title{log-likelihood function generator given data \code{x} for the weibull
distribution}
\usage{
weibull_fisher_info(x)
}
\arguments{
\item{x}{data}
}
\description{
log-likelihood function generator given data \code{x} for the weibull
distribution
}
