% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weibull.R
\name{weibull_shape_scale_fim}
\alias{weibull_shape_scale_fim}
\title{log-likelihood function generator given data \code{x} for the weibull
distribution}
\usage{
weibull_shape_scale_fim(x)
}
\arguments{
\item{x}{data}
}
\description{
log-likelihood function generator given data \code{x} for the weibull
distribution
}
