% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{aic.mle}
\alias{aic.mle}
\title{Method for obtaining the AIC of an \code{mle} object.}
\usage{
\method{aic}{mle}(x)
}
\arguments{
\item{x}{the \code{mle} object to obtain the AIC of}
}
\description{
Method for obtaining the AIC of an \code{mle} object.
}
