% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{aic}
\alias{aic}
\title{Generic method for obtaining the AIC of a fitted distribution object fit.}
\usage{
aic(x)
}
\arguments{
\item{x}{the object to obtain the AIC of}
}
\description{
Generic method for obtaining the AIC of a fitted distribution object fit.
}
