% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.mle.R
\docType{package}
\name{algebraic.mle}
\alias{algebraic.mle}
\title{algebraic.mle: A package for algebraically operating on maximum likelihood
               estimators.}
\description{
The algebraic.mle package provides three categories of important functions:
}
