% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.mle.R
\docType{package}
\name{algebraic.mle}
\alias{algebraic.mle}
\title{`algebraic.mle`: A package for algebraically operating on and generating
maximum likelihood estimators from existing maximum likelihood estimators.}
\description{
The object representing a fitted model is a type of `mle` object, the maximum
likelihood estimator of the model with respect to observed data.
}
\details{
It has a relatively rich API for working with these objects to help you
understand your MLE estimator.#'
}
