% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{bias.mle}
\alias{bias.mle}
\title{Computes the bias of an `mle` object assuming the large sample
approximation is valid and the MLE regularity conditions are satisfied.
In this case, the bias is zero (or zero vector).}
\usage{
\method{bias}{mle}(x, par = NULL, ...)
}
\arguments{
\item{x}{the `mle` object to compute the bias of.}

\item{par}{true parameter value. normally, unknown (NULL), in which case
we estimate the bias (say, using bootstrap)}

\item{...}{additional arguments to pass}
}
\description{
Computes the bias of an `mle` object assuming the large sample
approximation is valid and the MLE regularity conditions are satisfied.
In this case, the bias is zero (or zero vector).
}
