% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{bias.mle}
\alias{bias.mle}
\title{Computes the bias of an `mle` object assuming the large sample
approximation is valid and the MLE regularity conditions are satisfied.
In this case, the bias is zero (or zero vector).}
\usage{
\method{bias}{mle}(x, theta = NULL, ...)
}
\arguments{
\item{x}{the `mle` object to compute the bias of.}

\item{theta}{true parameter value. normally, unknown (NULL), in which case
we estimate the bias (say, using bootstrap)}

\item{...}{additional arguments to pass}
}
\value{
Numeric vector of zeros (asymptotic bias is zero under regularity
  conditions).
}
\description{
This is not a good estimate of the bias in general, but it's
arguably better than returning `NULL`.
}
