% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{bias.mle}
\alias{bias.mle}
\title{Computes the asymptotic bias of an \code{mle} object.}
\usage{
\method{bias}{mle}(x, ...)
}
\arguments{
\item{x}{the \code{mle} object to compute the bias of.}

\item{...}{pass additional arguments}
}
\description{
The bias of an estimator is just \code{E(point(mle)-theta)} where \code{theta}
is the true parameter value. Assuming the regularity conditions, the bias
is 0.
}
