% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{bias.mle_boot}
\alias{bias.mle_boot}
\title{Computes the estimate of the bias of a `mle_boot` object.}
\usage{
\method{bias}{mle_boot}(x, theta = NULL, ...)
}
\arguments{
\item{x}{the `mle_boot` object to compute the bias of.}

\item{theta}{true parameter value (not used for `mle_boot`).}

\item{...}{pass additional arguments (not used)}
}
\value{
Numeric vector of estimated bias (mean of bootstrap replicates minus
  original estimate).
}
\description{
Computes the estimate of the bias of a `mle_boot` object.
}
