% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{bias.mle_boot}
\alias{bias.mle_boot}
\title{Computes the estimate of the bias of a `mle_boot` object.}
\usage{
\method{bias}{mle_boot}(x, par = NULL, ...)
}
\arguments{
\item{x}{the `mle_boot` object to compute the bias of.}

\item{par}{for the `mle_boot`, we ignore this parameter.}

\item{...}{pass additional arguments (not used)}
}
\description{
Generally, we do not trust this to be a good estimate of the bias
of the MLE estimator, but it is still useful for comparing different
estimators.
}
