% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{bias.mle_exp}
\alias{bias.mle_exp}
\title{Computes the bias of an \code{exp_mle} object (exponential mle).}
\usage{
\method{bias}{mle_exp}(x, rate)
}
\arguments{
\item{x}{the \code{mle_exp} object to compute the bias of.}

\item{rate}{the true rate parameter value}
}
\description{
An unbiased estimator of the rate parameter of the exponential distribution
is given by: \code{1/(nobs(x)-1)*bias(x)}, where \code{x} is an
\code{mle_exp} object.
}
