% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{bias.mle_normal}
\alias{bias.mle_normal}
\title{Computes the bias of an `mle_normal` object.}
\usage{
\method{bias}{mle_normal}(x, par = NULL, ...)
}
\arguments{
\item{x}{An `mle_normal` object (subclass of `mle`) to compute the bias of.}

\item{par}{The true parameter value. If this is unknown (NULL), the bias is
estimated.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A numeric vector of length 2, the bias of the `mle_normal` estimator.
}
\description{
This function computes the bias of an `mle_normal` object.
It is a method for the generic function `bias`.
}
\seealso{
\code{\link{bias}} for the generic function.
}
