% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal_mu_var.R
\name{bias.mle_normal_mu_var}
\alias{bias.mle_normal_mu_var}
\title{Computes the bias of an \code{mle_normal} object.}
\usage{
\method{bias}{mle_normal_mu_var}(x, theta = NULL, ...)
}
\arguments{
\item{x}{the \code{mle} object to compute the bias of.}

\item{theta}{the true parameter value}

\item{...}{unused additional arguments}
}
\description{
Computes the bias of an \code{mle_normal} object.
}
