% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{bias}
\alias{bias}
\title{Computes the bias of an estimator object.}
\usage{
bias(x, theta)
}
\arguments{
\item{x}{the object to compute the bias of.}

\item{theta}{the true parameter value}
}
\description{
Computes the bias of an estimator object.
}
