% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{bias}
\alias{bias}
\title{Computes the bias of an estimator object.}
\usage{
bias(x, theta, ...)
}
\arguments{
\item{x}{the object to compute the bias of.}

\item{theta}{true parameter value. usually, this is unknown (NULL), in which
case we estimate the bias}

\item{...}{pass additional arguments}
}
\description{
Computes the bias of an estimator object.
}
