% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algeabraic.mle.R
\name{bias}
\alias{bias}
\title{Generic method for computing the bias of an estimator object.}
\usage{
bias(x, theta, ...)
}
\arguments{
\item{x}{the object to compute the bias of.}

\item{theta}{true parameter value. usually, this is unknown (NULL), in which
case we estimate the bias}

\item{...}{pass additional arguments}
}
\value{
The bias of the estimator. The return type depends on the specific
        method.
}
\description{
Generic method for computing the bias of an estimator object.
}
