% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{chisqr.test}
\alias{chisqr.test}
\title{chi-squar goodness-of-fit test}
\usage{
chisqr.test(obs, cdf, nbreaks = ceiling(sqrt(length(obs))), ...)
}
\arguments{
\item{obs}{a sample of observations}

\item{cdf}{the cdf of the reference distribution}

\item{nbreaks}{number of bins}

\item{...}{additional arguments to pass}
}
\value{
a hypothesis test object
}
\description{
chi-squar goodness-of-fit test
}
