% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{coef.mle_fit}
\alias{coef.mle_fit}
\title{Extract coefficients from an \code{mle_fit} object.}
\usage{
\method{coef}{mle_fit}(object, ...)
}
\arguments{
\item{object}{the \code{mle_fit} object}

\item{...}{additional arguments (not used)}
}
\value{
Named numeric vector of parameter estimates.
}
\description{
Delegates to \code{\link[algebraic.dist]{params}()}.
}
