% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_mles}
\alias{combine_mles}
\title{Internal workhorse for combine.}
\usage{
combine_mles(mles)
}
\arguments{
\item{mles}{A list of mle objects.}
}
\value{
An mle object.
}
\description{
Internal workhorse for combine.
}
\keyword{internal}
