% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{confint.mle}
\alias{confint.mle}
\title{Function to compute the confidence intervals of `mle` objects.}
\usage{
\method{confint}{mle}(x, level = 0.95, use_t_dist = TRUE, ...)
}
\arguments{
\item{x}{the `mle` object to compute the confidence intervals for}

\item{level}{confidence level, defaults to 0.95 (alpha=.05)}

\item{...}{additional arguments to pass}
}
\description{
Function to compute the confidence intervals of `mle` objects.
}
