% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{confint.mle}
\alias{confint.mle}
\title{Function to compute the confidence intervals of \code{mle} objects.}
\usage{
\method{confint}{mle}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the \code{mle} object to compute the confidence intervals for.}

\item{parm}{parameter indexes to compute the confidence intervals for,
defaults to all.}

\item{level}{confidence level, defaults to 0.95 (alpha=.05).}

\item{...}{additional arguments to pass.}
}
\description{
Function to compute the confidence intervals of \code{mle} objects.
}
