% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{confint.mle_boot}
\alias{confint.mle_boot}
\title{Method for obtained the confidence interval of an `mle_boot` object.
Note: This impelements the `vcov` method defined in `stats`.}
\usage{
\method{confint}{mle_boot}(
  object,
  parm = NULL,
  level = 0.95,
  type = c("norm", "basic", "perc", "bca"),
  ...
)
}
\arguments{
\item{object}{the `mle_boot` object to obtain the confidence interval of}

\item{parm}{the parameter to obtain the confidence interval of (not used)}

\item{level}{the confidence level}

\item{type}{the type of confidence interval to compute}

\item{...}{additional arguments to pass into `boot.ci`}
}
\value{
Matrix of bootstrap confidence intervals with columns for lower and
  upper bounds.
}
\description{
Method for obtained the confidence interval of an `mle_boot` object.
Note: This impelements the `vcov` method defined in `stats`.
}
