% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint_from_sigma}
\alias{confint_from_sigma}
\title{Function to compute the confidence intervals from a variance-covariance matrix}
\usage{
confint_from_sigma(sigma, theta, level = 0.95)
}
\arguments{
\item{sigma}{either the variance-covariance matrix or the vector of variances
of the parameter estimator}

\item{theta}{the point estimate}

\item{level}{confidence level, defaults to 0.95 (alpha=.05)}
}
\description{
Function to compute the confidence intervals from a variance-covariance matrix
}
