% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{dim.mle_fit}
\alias{dim.mle_fit}
\title{Dimension (number of parameters) of an MLE.}
\usage{
\method{dim}{mle_fit}(x)
}
\arguments{
\item{x}{An \code{mle_fit} object.}
}
\value{
Integer; the number of parameters.
}
\description{
Dimension (number of parameters) of an MLE.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), loglike = -120, nobs = 100L)
dim(fit)
}
