% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{exp_loglike}
\alias{exp_loglike}
\title{log-likelihood function generator given data \code{x} for the exponential
distribution}
\usage{
exp_loglike(x)
}
\arguments{
\item{x}{data}
}
\description{
log-likelihood function generator given data \code{x} for the exponential
distribution
}
