% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{fisher_info}
\alias{fisher_info}
\title{Generic method for obtaining the fisher information
matrix of an \code{mle} object.}
\usage{
fisher_info(x, ...)
}
\arguments{
\item{x}{the object to obtain the fisher information of}

\item{...}{additional arguments to pass}
}
\description{
Generic method for obtaining the fisher information
matrix of an \code{mle} object.
}
