% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{is_mle_boot}
\alias{is_mle_boot}
\title{Determine if an object is an `mle_boot` object.}
\usage{
is_mle_boot(x)
}
\arguments{
\item{x}{the object to test}
}
\description{
Determine if an object is an `mle_boot` object.
}
