% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{\code{mle} makes an \code{mle} object.}
\usage{
mle(
  theta.hat,
  loglike = NULL,
  score = NULL,
  sigma = NULL,
  info = NULL,
  obs = NULL,
  nobs = NULL,
  superclasses = NULL
)
}
\arguments{
\item{theta.hat}{the MLE}

\item{loglike}{the log-likelihood of \code{theta.hat} given the data}

\item{score}{the score function evaluated at \code{theta.hat}}

\item{sigma}{the variance-covariance matrix of \code{theta.hat} given that data}

\item{info}{the information matrix of \code{theta.hat} given the data}

\item{obs}{observation (sample) data}

\item{nobs}{number of observations in \code{obs}}

\item{superclasses}{class (or classes) with \code{mle} as base}
}
\description{
\code{mle} makes an \code{mle} object.
}
