% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{mle_boot}
\alias{mle_boot}
\title{A function for estimating the empirical sampling distribution the
MLE given a MLE solver and a sample using the Bootstrap method.}
\usage{
mle_boot(mle_solver, data, R, ...)
}
\arguments{
\item{mle_solver}{given a data, find the MLE.}

\item{data}{data for resampling, where for each resample we generate an MLE}

\item{R}{bootstrap replicates}

\item{...}{additional arguments to pass.}
}
\description{
A function for estimating the empirical sampling distribution the
MLE given a MLE solver and a sample using the Bootstrap method.
}
