% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{mle_boot}
\alias{mle_boot}
\title{Bootstrap MLE}
\usage{
mle_boot(mle_solver, data, R = 999, ...)
}
\arguments{
\item{mle_solver}{given data, find the MLE.}

\item{data}{data for resampling, where for each resample we generate an MLE}

\item{R}{bootstrap replicates, defaults to 999}

\item{...}{additional arguments to pass.}
}
\value{
an `mle_boot` object, which is an `mle` object with a `boot` object
        as its parent.
}
\description{
Sometimes, the large sample asymptotic theory of MLEs is not applicable.
In such cases, we can use the bootstrap to estimate the sampling distribution
of the MLE.
}
\details{
Look up the `boot` package for more information on the bootstrap and
how to use it. You can pass additional arguments to the `boot` function
using the `...` argument to `mle_boot`.
}
\examples{
data <- rnorm(15)
solver <- function(data, ind) {
    point(mle_normal(data[ind]))
}
theta.bt <- mle_boot(solver, data)
}
