% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{mle_boot_loglike}
\alias{mle_boot_loglike}
\title{A function for computing the sampling distribution of a statistic of the
MLE's sampling distribution using the Bootstrap method.}
\usage{
mle_boot_loglike(
  mle,
  loglike.gen,
  data = NULL,
  R = NULL,
  method = mle_newton_raphson,
  ...
)
}
\arguments{
\item{mle}{a fitted \code{mle} object.}

\item{loglike.gen}{a generator for the log-likelihood function; it accepts
observations and constructs the log-likelihood function}

\item{data}{data for generating MLEs for the bootstrap resampling.}

\item{R}{bootstrap replicates}

\item{method}{method for solving the MLE, defaults to numerically solving
the root of the gradient of the log-likelihood using
Newton-raphson.}

\item{...}{additional arguments to pass.}
}
\description{
A function for computing the sampling distribution of a statistic of the
MLE's sampling distribution using the Bootstrap method.
}
