% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{mle_exp}
\alias{mle_exp}
\title{MLE of the rate parameter of the exponential distribution given a
random sample of observations drawn from it.}
\usage{
mle_exp(x)
}
\arguments{
\item{x}{vector of draws from exponential distribution.}
}
\value{
an \code{mle_exp} object.
}
\description{
Of course, the draws are unlikely to be exponential, but the exponential may
be a sufficiently good model. Hypothesis testing, such as relative
likelihoods, can be used to assess the appropriateness of the exponential
model to the data.
}
