% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{mle_exp}
\alias{mle_exp}
\title{MLE of the rate parameter when we assume the sample is i.i.d. and drawn
from the exponential distribution.}
\usage{
mle_exp(x)
}
\arguments{
\item{x}{a sample of observations}
}
\value{
an \code{mle} object.
}
\description{
Of course, the draws are unlikely to be exponential, but the exponential may
be a sufficiently good model. Hypothesis testing, such as relative
likelihoods, can be used to assess the appropriateness of the exponential
model to the data.
}
