% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_numerical.R
\name{mle_gradient_ascent}
\alias{mle_gradient_ascent}
\title{Iterative MLE method using gradient ascent}
\usage{
mle_gradient_ascent(
  l,
  theta0,
  score = function(theta) numDeriv::grad(l, theta),
  stop_cond = function(theta1, theta0) abs(max(theta1 - theta0)) < 1e-04,
  sup = function(theta) all(theta > 0),
  r = 0.75,
  max_iter = 0L
)
}
\arguments{
\item{l}{log-likelihood function}

\item{theta0}{initial guess of theta with \eqn{p} components}

\item{score}{score function of type \eqn{R^p -> R^p}}

\item{stop_cond}{stopping condition function of type \eqn{R^p \times R^p \mapsto \{true,false\}}}

\item{sup}{domain of support function of type \eqn{R^p \mapsto \{true,false\}} for the log-likelhood function \eqn{l}}

\item{r}{backing tracking parameter}

\item{max_iter}{maximum iterations}
}
\description{
Iterative MLE method using gradient ascent
}
