% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_numerical.R
\name{mle_gradient_ascent}
\alias{mle_gradient_ascent}
\title{Iterative MLE method using gradient ascent}
\usage{
mle_gradient_ascent(
  l,
  theta0,
  score,
  eps = 1e-05,
  sup = NULL,
  eta = 1,
  r = 0.5,
  max_iter = 0L,
  debug = F
)
}
\arguments{
\item{l}{log-likelihood function}

\item{theta0}{initial guess of theta with \eqn{p} components}

\item{score}{score function of type \eqn{R^p -> R^p}}

\item{eps}{stopping condition}

\item{sup}{domain of support function of type \eqn{R^p \mapsto \{true,false\}} for the log-likelihood function \eqn{l}}

\item{eta}{learning rate, defaults to 1}

\item{r}{backing tracking parameter}

\item{max_iter}{maximum iterations}

\item{debug}{Boolean, output debugging information if TRUE; defaults to FALSE}
}
\description{
Iterative MLE method using gradient ascent
}
