% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_numerical.R
\name{mle_iterative}
\alias{mle_iterative}
\title{General iterative MLE method}
\usage{
mle_iterative(
  l,
  theta0,
  dir = function(theta) grad(l, theta),
  stop_cond = function(theta1, theta0) abs(max(theta1 - theta0)) < 1e-04,
  sup = function(theta) all(theta > 0),
  eta = 1,
  r = 0.5,
  max_iter = 0L
)
}
\arguments{
\item{l}{log-likelihood function of type \eqn{R^p \mapsto R}}

\item{theta0}{initial guess of \eqn{\theta} with \eqn{p} components}

\item{dir}{promising direction function of type \eqn{R^p \mapsto R^p}}

\item{stop_cond}{stopping condition function of type \eqn{R^p \times R^p \mapsto \{true,false\}}}

\item{sup}{domain of support function of type \eqn{R^p \mapsto \{true,false\}} for the log-likelhood function \eqn{l}}

\item{eta}{learning rate}

\item{r}{backing tracking parameter}

\item{max_iter}{maximum number of iterations}
}
\description{
General iterative MLE method
}
