% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_mle.R
\name{mle_local_search}
\alias{mle_local_search}
\title{mle_local_search}
\usage{
mle_local_search(ll, dir, theta0, options = list())
}
\arguments{
\item{ll}{function, log-likelihood function}

\item{dir}{function, promising direction function}

\item{theta0}{numeric, initial guess}

\item{options}{list, options for the local search, see function description.}
}
\value{
an `mle` object with additional attributes `iter` and `converged`
        and optionally `path` if `trace` is TRUE.
}
\description{
This assumes the MLE is an interior point and that you have provided
an initial guess `theta0` that is near it. Use a global search method
like `sim_anneal` to find a good initial guess.
}
\section{Functions}{
\itemize{
\item \code{mle_local_search()}: options

}}
\section{Fields}{

\describe{
\item{\code{sup}}{function, domain of support for log-likelihood}

\item{\code{eta}}{numeric, learning rate, defaults to 1}

\item{\code{max_iter}}{integer, maximum number of iterations, defaults to 1000}

\item{\code{max_iter_ls}}{integer, maximum number of iterations for the
line search, defaults to 1000}

\item{\code{abs_tol}}{numeric, tolerance for convergence, defaults to NULL
(use rel_tol instead)}

\item{\code{rel_tol}}{numeric, relative tolerance for convergence, defaults to 1e-5}

\item{\code{r}}{numeric, backtracking line search parameter, defaults to 0.8}

\item{\code{proj}}{function, projection function to enforce domain of support}

\item{\code{norm}}{function, we pass the difference of successive theta updates
to the norm to check for convergence, defaults to the infinity norm.}

\item{\code{debug}}{logical, output debugging information if TRUE; default FALSE}

\item{\code{trace}}{logical, if TRUE store the path of the search in the `path`
attribute of the output; default FALSE}
}}

