% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_numerical.R
\name{mle_newton_raphson}
\alias{mle_newton_raphson}
\title{Iterative MLE method using the Newton-Raphson method}
\usage{
mle_newton_raphson(
  l,
  theta0,
  info,
  score,
  stop_cond = function(theta1, theta0) abs(max(theta1 - theta0)) < 1e-04,
  sup = function(theta) all(theta > 0),
  r = 0.75,
  max_iter = 0L
)
}
\arguments{
\item{l}{log-likelihood function of type \code{R^p -> R}}

\item{theta0}{initial guess of theta with \code{p} components}

\item{info}{information matrix function of type \code{R^p -> R^{p-by-q}}}

\item{score}{score function of type \code{R^p -> R^p}}

\item{stop_cond}{stopping condition function of type \code{(R^p,R^p) -> \{true,false\}}}

\item{sup}{domain of support function of type \code{R^p -> \{true,false\}} for the log-likelhood function \code{l}}

\item{r}{backing tracking parameter}

\item{max_iter}{maximum number of iterations}
}
\description{
Iterative MLE method using the Newton-Raphson method
}
