% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_mle.R
\name{mle_newton_raphson}
\alias{mle_newton_raphson}
\title{MLE method using the Newton-Raphson method}
\usage{
mle_newton_raphson(
  ll,
  theta0,
  info,
  score,
  eps = 1e-05,
  sup = NULL,
  eta = 1,
  max_iter = 0L,
  r = 0.5,
  proj = NULL,
  tol = NULL,
  debug = FALSE,
  inverted = FALSE
)
}
\arguments{
\item{ll}{log-likelihood function of type `R^p -> R`}

\item{theta0}{initial guess of theta with `p` components}

\item{info}{FIM function of type `R^p -> R^{p-by-q}`.}

\item{score}{score Function of type `R^p -> R^p`, the gradient of `ll``}

\item{eps}{numeric, stopping criterion}

\item{sup}{domain of support function of type `R^p -> {T,F}` for `ll`}

\item{eta}{numeric, learning rate, defaults to 1}

\item{max_iter}{maximum number of iterations}

\item{r}{backing tracking parameter}

\item{debug}{logical, output debugging info if `TRUE`; defaults to `FALSE`}

\item{inverted}{if `TRUE` `info` is covariance instead of FIM}
}
\description{
MLE method using the Newton-Raphson method
}
