% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_mle.R
\name{mle_newton_raphson}
\alias{mle_newton_raphson}
\title{mle_newton_raphson}
\usage{
mle_newton_raphson(ll, score, info, theta0, inverted = FALSE, options = list())
}
\arguments{
\item{ll}{log-likelihood function}

\item{score}{score Function, the gradient of log-likelihood}

\item{info}{FIM function}

\item{theta0}{initial guess}

\item{inverted}{logical, if TRUE `info` is covariance instead of FIM}

\item{options}{list, options for the local search, see `mle_local_search`}
}
\value{
an object of class `mle_newton_raphson`, which is an `mle` object
}
\description{
MLE method using the Newton-Raphson method
}
