% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_numerical.R
\name{mle_newton_raphson}
\alias{mle_newton_raphson}
\title{Iterative MLE method using the Newton-Raphson method}
\usage{
mle_newton_raphson(
  l,
  theta0,
  info,
  score,
  stop_cond = NULL,
  sup = NULL,
  eta = 1,
  r = 0.5,
  max_iter = 0L
)
}
\arguments{
\item{l}{log-likelihood function of type \code{R^p -> R}}

\item{theta0}{initial guess of theta with \code{p} components}

\item{info}{information matrix function of type \code{R^p -> R^{p-by-q}}}

\item{score}{score function of type \code{R^p -> R^p}}

\item{stop_cond}{stopping condition function of type \code{(R^p,R^p) -> \{T,F\}}}

\item{sup}{domain of support function of type \code{R^p -> \{T,F\}} for the log-likelihood function \code{l}}

\item{eta}{learning rate, defaults to 1}

\item{r}{backing tracking parameter}

\item{max_iter}{maximum number of iterations}
}
\description{
Iterative MLE method using the Newton-Raphson method
}
