% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{mle_normal}
\alias{mle_normal}
\title{MLE of the (mu,var) parameter vector when we assume the sample is i.i.d. and
drawn from the normal distribution.}
\usage{
mle_normal(x, keep_obs = TRUE)
}
\arguments{
\item{x}{a sample of observations}

\item{keep_obs}{whether to store observations in `mle_normal` object}
}
\value{
an `mle` object.
}
\description{
Of course, the draws are unlikely to be normal, but the normal distribution
is often a good model. Hypothesis testing, such as relative
likelihoods, can be used to assess the appropriateness of the normal
model to the data.
}
