% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{mle_normal_mu_var}
\alias{mle_normal_mu_var}
\title{MLE of the (mu,sigma) parameter vector when we assume the sample is i.i.d. and
drawn from the normal distribution.}
\usage{
mle_normal_mu_var(x)
}
\arguments{
\item{x}{a sample of observations}
}
\description{
MLE of the (mu,sigma) parameter vector when we assume the sample is i.i.d. and
drawn from the normal distribution.
}
