% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_numerical.R
\name{mle_numerical}
\alias{mle_numerical}
\title{mle_optim}
\usage{
mle_numerical(sol, options = list())
}
\arguments{
\item{sol}{the output of `optim` or `newton_raphson`}

\item{options}{list, options for things like sigma and FIM}
}
\value{
a `numerical_mle` object.
}
\description{
This function takes the output of `optim`, `newton_raphson`, or `sim_anneal`
and turns it into an `mle_numerical` (subclass of `mle`) object.
}
