% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weibull.R
\name{mle_weibull}
\alias{mle_weibull}
\title{MLE of the rate parameter when we assume the sample is i.i.d. and drawn
from the weibull distribution.}
\usage{
mle_weibull(
  x,
  par0 = c(1, 1),
  pgtol = 1e-07,
  maxit = 10000L,
  keep_obs = FALSE,
  ...
)
}
\arguments{
\item{x}{a sample of observations}

\item{par0}{initial estimate of the parameters, defaults to `c(1,1)`}

\item{pgtol}{the convergence tolerance for the optimization algorithm}

\item{keep_obs}{Boolean, specifies whether to keep observations}
}
\value{
an `mle` object.
}
\description{
MLE of the rate parameter when we assume the sample is i.i.d. and drawn
from the weibull distribution.
}
