% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weibull.R
\name{mle_weibull}
\alias{mle_weibull}
\title{MLE of the rate parameter when we assume the sample is i.i.d. and drawn
from the weibull distribution.}
\usage{
mle_weibull(x, eps = 1e-05)
}
\arguments{
\item{x}{a sample of observations}
}
\value{
an \code{mle} object.
}
\description{
MLE of the rate parameter when we assume the sample is i.i.d. and drawn
from the weibull distribution.
}
