% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weighted.R
\name{mle_weighted}
\alias{mle_weighted}
\title{mle_weighted takes a list of MLEs for some parameter gamma and then
combines them to form the MLE for gamma over all the information in the
samples used to generate the list of MLEs.}
\usage{
mle_weighted(mles)
}
\arguments{
\item{mles}{A list of MLEs, all for the same parameter.}
}
\description{
mle_weighted takes a list of MLEs for some parameter gamma and then
combines them to form the MLE for gamma over all the information in the
samples used to generate the list of MLEs.
}
