% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weighted.R
\name{mle_weighted}
\alias{mle_weighted}
\title{Accepts a list of `mle` objects for some parameter, say `theta`,
and combines them into a single estimator `mle_weighted`.}
\usage{
mle_weighted(mles)
}
\arguments{
\item{mles}{A list of `mle` objects, all for the same parameter.}
}
\value{
An object of type `mle_weighted` (which inherits from
        `mle`) that is the weighted sum of the `mle` objects.
}
\description{
It combines the `mle` objects by adding them together, weighted by
the inverse of their respective variance-covariance matrix (information
matrix). Intuitively, the higher the variance, the less weight an `mle`
is given in the summation, or alternatively, the more information it
has about the parameter, the more weight it is given in the summation.
}
\details{
Each `mle` object should have an `observed_fim` method, which returns
the Fisher information matrix (FIM) for the parameter. The FIM is
assumed to be the negative of the expected value of the Hessian of the
log-likelihood function. The `mle` objects should also have a `params`
method, which returns the parameter vector.

We assume that the observations used to estimate each of the MLE objects
in `mles` are independent.
}
