% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mse.mle}
\alias{mse.mle}
\title{Computes the asymptotic MSE of an \code{mle} object.}
\usage{
\method{mse}{mle}(x, ...)
}
\arguments{
\item{x}{the \code{mle} object to compute the MSE of.}

\item{...}{pass additional arguments}
}
\description{
The MSE of an estimator is just the expected sum of squared differences,
e.g., if the true parameter value is \code{x} and we have an estimator \code{x.hat},
then the MSE is
\code{mse(x.hat) = E[(x.hat-x)^2] = trace(vcov(x.hat)) + (bias(x.hat))^2}.
}
\details{
Since \code{x} is not typically known, the bias is not a statistic, and thus the
MSE for biased estimators is not typically a statistic that can be computed.
However, assuming the regularity conditions, the MLE is asymptotically
unbiased, and thus for sufficiently large samples,
\code{mse(x.hat) = trace(vcov(theta.hat))}.
}
