% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{mse.mle_boot}
\alias{mse.mle_boot}
\title{Computes the estimate of the MSE of a `boot` object.}
\usage{
\method{mse}{mle_boot}(x, par = NULL, ...)
}
\arguments{
\item{x}{the `boot` object to compute the MSE of.}

\item{par}{for the `mle_boot`, we ignore this parameter}

\item{...}{pass additional arguments (not used)}
}
\description{
Computes the estimate of the MSE of a `boot` object.
}
