% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{mse}
\alias{mse}
\title{Generic function for obtaining the mean squared error (MSE) of an estimator,
`mse(x) = E[(x-mu)^2]` where `mu` is the true parameter value.}
\usage{
mse(x, theta)
}
\arguments{
\item{x}{the object to compute the MSE of}

\item{theta}{the true parameter value}
}
\description{
Generic function for obtaining the mean squared error (MSE) of an estimator,
`mse(x) = E[(x-mu)^2]` where `mu` is the true parameter value.
}
