% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{nobs.mle}
\alias{nobs.mle}
\title{Method for obtaining the number of observations in the sample used by
an \code{mle}.}
\usage{
\method{nobs}{mle}(x)
}
\arguments{
\item{x}{the \code{mle} object to obtain the number of observations for}
}
\description{
Method for obtaining the number of observations in the sample used by
an \code{mle}.
}
