% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{nobs.mle}
\alias{nobs.mle}
\title{Method for obtaining the number of observations in the sample used by
an `mle`.}
\usage{
\method{nobs}{mle}(object, ...)
}
\arguments{
\item{object}{the `mle` object to obtain the number of observations for}

\item{...}{additional arguments to pass (not used)}
}
\description{
Method for obtaining the number of observations in the sample used by
an `mle`.
}
