% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{nobs.mle_boot}
\alias{nobs.mle_boot}
\title{Method for obtaining the number of observations in the sample used by
an `mle`.}
\usage{
\method{nobs}{mle_boot}(object, ...)
}
\arguments{
\item{object}{the `mle` object to obtain the number of observations for}

\item{...}{additional arguments to pass (not used)}
}
\value{
Integer number of observations in the original sample.
}
\description{
Method for obtaining the number of observations in the sample used by
an `mle`.
}
