% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{nobs.mle_fit}
\alias{nobs.mle_fit}
\title{Method for obtaining the number of observations in the sample used by
an `mle_fit`.}
\usage{
\method{nobs}{mle_fit}(object, ...)
}
\arguments{
\item{object}{the `mle_fit` object to obtain the number of observations for}

\item{...}{additional arguments to pass (not used)}
}
\value{
Integer number of observations, or NULL if not available.
}
\description{
Method for obtaining the number of observations in the sample used by
an `mle_fit`.
}
