% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{nparams.mle_fit}
\alias{nparams.mle_fit}
\title{Method for obtaining the number of parameters of an `mle_fit` object.}
\usage{
\method{nparams}{mle_fit}(x)
}
\arguments{
\item{x}{the `mle_fit` object to obtain the number of parameters of}
}
\value{
Integer number of parameters.
}
\description{
Method for obtaining the number of parameters of an `mle_fit` object.
}
