% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{nparams}
\alias{nparams}
\title{Generic method for obtaining the number of parameters of an MLE fit.}
\usage{
nparams(x)
}
\arguments{
\item{x}{the object to obtain the number of parameters for}
}
\description{
Generic method for obtaining the number of parameters of an MLE fit.
}
