% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{obs.mle_boot}
\alias{obs.mle_boot}
\title{Method for obtaining the observations used by the `mle`.}
\usage{
\method{obs}{mle_boot}(x)
}
\arguments{
\item{x}{the `mle` object to obtain the number of observations for}
}
\value{
The original data used for bootstrapping.
}
\description{
Method for obtaining the observations used by the `mle`.
}
