% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{obs.mle_fit}
\alias{obs.mle_fit}
\title{Method for obtaining the observations used by the `mle_fit` object `x`.}
\usage{
\method{obs}{mle_fit}(x)
}
\arguments{
\item{x}{the `mle_fit` object to obtain the number of observations for}
}
\value{
The observation data used to fit the MLE, or NULL if not stored.
}
\description{
Method for obtaining the observations used by the `mle_fit` object `x`.
}
