% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{obs}
\alias{obs}
\title{Generic function for obtaining the observations used by a fitted model `object`.}
\usage{
obs(object, ...)
}
\arguments{
\item{object}{the fitted object to obtain the number of observations used by the fit}

\item{...}{additional arguments to pass}
}
\description{
Generic function for obtaining the observations used by a fitted model `object`.
}
