% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{orthogonal}
\alias{orthogonal}
\title{Method for determining the orthogonal parameters of an estimator.}
\usage{
orthogonal(x)
}
\arguments{
\item{x}{the estimator}
}
\description{
Method for determining the orthogonal parameters of an estimator.
}
