% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algeabraic.mle.R
\name{orthogonal}
\alias{orthogonal}
\title{Method for determining the orthogonal parameters of an estimator.}
\usage{
orthogonal(x, tol, ...)
}
\arguments{
\item{x}{the estimator}

\item{tol}{the tolerance for determining if a number is close enough to zero}

\item{...}{additional arguments to pass}
}
\description{
Method for determining the orthogonal parameters of an estimator.
}
