% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{params.mle}
\alias{params.mle}
\title{Method for obtaining the parameters of an `mle` object.}
\usage{
\method{params}{mle}(x)
}
\arguments{
\item{x}{the `mle` object to obtain the parameters of}
}
\description{
Method for obtaining the parameters of an `mle` object.
}
