% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_boot.R
\name{params.mle_boot}
\alias{params.mle_boot}
\title{Method for obtaining the parameters of an `boot` object.}
\usage{
\method{params}{mle_boot}(x)
}
\arguments{
\item{x}{the `boot` object to obtain the parameters of.}
}
\value{
Numeric vector of parameter estimates (the original MLE).
}
\description{
Method for obtaining the parameters of an `boot` object.
}
