% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{params}
\alias{params}
\title{Method for obtaining the parameters of a fitted object.}
\usage{
params(x)
}
\arguments{
\item{x}{the fitted object to obtain the parameters of}
}
\description{
Method for obtaining the parameters of a fitted object.
}
