% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{point.boot}
\alias{point.boot}
\title{Method for obtaining the parameter estimate of a \code{boot} object.}
\usage{
\method{point}{boot}(x, ...)

\method{point}{boot}(x, ...)
}
\arguments{
\item{x}{the \code{boot} object.}

\item{...}{pass additional arguments}
}
\description{
Method for obtaining the parameter estimate of a \code{boot} object.

Computes the point estimate of an \code{mle} object.
}
