% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{point.mle}
\alias{point.mle}
\title{Computes the point estimate of an \code{mle} object.}
\usage{
\method{point}{mle}(x, ...)
}
\arguments{
\item{x}{the \code{mle} object.}

\item{...}{pass additional arguments}
}
\description{
Computes the point estimate of an \code{mle} object.
}
