% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{pred.mle}
\alias{pred.mle}
\title{Predictive pdf of T given mle.hat estimate for the parameters of T}
\usage{
\method{pred}{mle}(x, samp, alpha = 0.05)
}
\arguments{
\item{x}{\code{mle} object}

\item{samp}{\code{mle} sampler for parametric distribution for the mle \code{x}}

\item{alpha}{(1-alpha)-predictive interval for T|x}
}
\description{
Let
  T|theta.hat ~ f(theta.hat)
be the pdf of T given parameter value theta.hat and
  theta.hat ~ normal(mean(theta.hat),vcov(theta.hat))
be the mvn of the MLE for theta.hat.
}
\details{
Then,
  (T,theta.hat) ~ f(t,theta.hat) = f(t|theta.hat) f(theta.hat)
is the joint distribution of (T,theta.hat). To find f(t), we integrate
f(t,theta.hat) over theta.hat.
}
