% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Function for obtaining a prediction for a function \code{f} with
respect to an \code{mle} object \code{x}.}
\usage{
pred(x, f, N, p = 0.95, D = function(x) x)
}
\arguments{
\item{x}{the \code{mle} object}

\item{f}{the function to predict the value of with respect to input \code{x}}

\item{N}{the sample size}

\item{p}{the value that defines the confidence region}

\item{D}{the distance measure, defaults to identity function}
}
\description{
We use Monte-Carlo simulation to get \code{N} sample points within the
\code{p}-confidence region and taking the maximum value with respect to
some distance measure.
}
