% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{print.mle}
\alias{print.mle}
\title{Method for obtaining the number of observations in the sample used by
an `mle` object `x`.}
\usage{
\method{print}{mle}(x, ...)
}
\arguments{
\item{x}{the `mle` object to print}

\item{...}{additional arguments to pass}
}
\description{
Method for obtaining the number of observations in the sample used by
an `mle` object `x`.
}
