% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{print.mle_fit}
\alias{print.mle_fit}
\title{Print method for `mle_fit` objects.}
\usage{
\method{print}{mle_fit}(x, ...)
}
\arguments{
\item{x}{the `mle_fit` object to print}

\item{...}{additional arguments to pass}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for `mle_fit` objects.
}
