% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{print.summary_mle_fit}
\alias{print.summary_mle_fit}
\title{Function for printing a `summary` object for an `mle_fit` object.}
\usage{
\method{print}{summary_mle_fit}(x, ...)
}
\arguments{
\item{x}{the `summary_mle_fit` object}

\item{...}{pass additional arguments}
}
\value{
Invisibly returns \code{x}.
}
\description{
Function for printing a `summary` object for an `mle_fit` object.
}
