% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_map.R
\name{rmap.mle}
\alias{rmap.mle}
\title{\code{rmap.mle} computes the distribution of \code{f(x)} as a function of
\code{mle} object \code{x}.}
\usage{
\method{rmap}{mle}(x, f, n = 1000, ...)
}
\arguments{
\item{x}{a list of \code{mle} objects.}

\item{f}{a function that accepts objects like x (e.g., a vector).}

\item{n}{number of samples to take from \code{x} to estimate distribution
of \code{f(x)}.}

\item{...}{additional arguments to pass to the \code{mle} sampler.}
}
\description{
When we apply \code{f} to an argument \code{x}, denoted by \code{f(x)}, we
are evaluating \code{f} by substituting the formal parameters of \code{f}
with the argument \code{x}. However, if the argument is a random variable,
then \code{f(x)} is also a random variable.
}
\details{
By the invariance property of the MLE, if \code{x} is an \code{mle} object,
asymptotically, \code{f(x)} is normally distributed with an approximation
given by \code{f(x) ~ normal(f(point(x)),sigma)} where \code{sigma} is an
estimate of the variance-covariance of \code{f(x)}, e.g., the sample
covariance of \code{f(x1),...f(xn)} where \code{xj} is sampled from
\code{sampler(x)}.
}
