% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{rmap}
\alias{rmap}
\title{Generic function for computing the distribution of \code{f(x)} where \code{x}
models a \code{dist} (distribution) object.}
\usage{
rmap(x, f, ...)
}
\arguments{
\item{x}{a list of \code{dist} (distrubtion) objects.}

\item{f}{a function that accepts arguments sampled from the \code{dist} items
in \code{x}.}

\item{...}{additional arguments to pass.}
}
\description{
When we apply a function \code{f} to an argument \code{x}, we are applying
the function to the argument, or evaluating the function by substituting the
formal parameters of the function with the arguments. However, if the
argument is a random variable, then \code{f(x)} is also a random variable.
}
