% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{rmap}
\alias{rmap}
\title{Generic function for computing the distribution of `f(x)` where `x`
models a distribution (random element) object.}
\usage{
rmap(x, g, ...)
}
\arguments{
\item{x}{a list of distribution (random element) objects.}

\item{g}{a function that accepts arguments sampled from the distribution
objects in `x`.}

\item{...}{additional arguments to pass.}
}
\description{
When we apply a function `f` to a distribution `x`, the result
is a distribution object `f(x)`.
}
