% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{sampler.mle_fit}
\alias{sampler.mle_fit}
\title{Method for sampling from an `mle_fit` object.}
\usage{
\method{sampler}{mle_fit}(x, ...)
}
\arguments{
\item{x}{the `mle_fit` object to create sampler for}

\item{...}{additional arguments to pass}
}
\value{
A function that takes parameter \code{n} and returns \code{n} samples
  from the asymptotic distribution of the MLE.
}
\description{
It creates a sampler for the `mle_fit` object. It returns a function
that accepts a single parameter `n` denoting the number of samples
to draw from the `mle_fit` object.
}
