% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{sampler}
\alias{sampler}
\title{Generic function for sampling from distributions.}
\usage{
sampler(x, ...)
}
\arguments{
\item{x}{the \code{x} object to create sampler for}

\item{...}{additional arguments to pass.}
}
\description{
It creates a sampler for the \code{x} object. It returns a function
that accepts a single parameter \code{n} denoting the number of samples
to draw from the \code{x} object.
}
