% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algeabraic.mle.R
\name{score_val}
\alias{score_val}
\title{score}
\usage{
score_val(x, ...)
}
\arguments{
\item{x}{the object to compute the score of.}

\item{...}{pass additional arguments}
}
\description{
Generic function for computing the score of an estimator
object with respect to its log-likelihood.
}
