% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algeabraic.mle.R
\name{score_val}
\alias{score_val}
\title{Generic method for computing the score of an estimator
object (gradient of its log-likelihood function evaluated
at the MLE).}
\usage{
score_val(x, ...)
}
\arguments{
\item{x}{the object to compute the score of.}

\item{...}{pass additional arguments}
}
\description{
Generic method for computing the score of an estimator
object (gradient of its log-likelihood function evaluated
at the MLE).
}
