% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{se.mle}
\alias{se.mle}
\title{Function for obtaining an estimate of the standard error of the MLE
`object`.}
\usage{
\method{se}{mle}(object, ...)
}
\arguments{
\item{object}{the MLE object}

\item{...}{pass additional arguments}
}
\description{
Function for obtaining an estimate of the standard error of the MLE
`object`.
}
