% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algeabraic.mle.R
\name{se}
\alias{se}
\title{Method for obtaining the standard error of an estimator.}
\usage{
se(x, ...)
}
\arguments{
\item{x}{the estimator}

\item{...}{additional arguments to pass}
}
\description{
Method for obtaining the standard error of an estimator.
}
