% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.mle.R
\name{se}
\alias{se}
\title{Generic method for obtaining the standard errors of an estimator.}
\usage{
se(x, ...)
}
\arguments{
\item{x}{the estimator}

\item{...}{additional arguments to pass}
}
\value{
Vector of standard errors for each parameter.
}
\description{
Generic method for obtaining the standard errors of an estimator.
}
