% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{se}
\alias{se}
\title{Method for obtaining the standard error of an estimator.}
\usage{
se(object)
}
\arguments{
\item{object}{the estimator}
}
\description{
Method for obtaining the standard error of an estimator.
}
