% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_anneal}
\alias{sim_anneal}
\title{Simulated Annealing}
\usage{
sim_anneal(
  f,
  x0,
  t_init = 100,
  t_end = 0.001,
  alpha = 0.99,
  sup = function(x) TRUE,
  neigh = function(x) x + rnorm(length(x)),
  max_iter = NULL
)
}
\arguments{
\item{f}{Objective function to maximize}

\item{x0}{Initial guess}

\item{t_init}{Initial temperature}

\item{t_end}{Final temperature}

\item{alpha}{Cooling factor}

\item{sup}{Support function, returns TRUE if x is in the domain of f}

\item{neigh}{Neighborhood function, returns a random neighbor of x}

\item{max_iter}{Maximum number of iterations, used instead of t_end
if not NULL, defaults to NULL}
}
\value{
list with best solution (argmax) and its corresponding
        objective function value (max)
}
\description{
Simulated Annealing
}
