% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{summary.mle}
\alias{summary.mle}
\title{Function for obtaining a summary of `object`, which is a fitted
`mle` object.}
\usage{
\method{summary}{mle}(object, ...)
}
\arguments{
\item{object}{the `mle` object}

\item{...}{pass additional arguments}
}
\description{
Function for obtaining a summary of `object`, which is a fitted
`mle` object.
}
