% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{vcov.mle}
\alias{vcov.mle}
\title{Computes the variance-covariance matrix of the MLE.}
\usage{
\method{vcov}{mle}(object, ...)
}
\arguments{
\item{object}{a fitted \code{mle} object.}

\item{...}{this is not used for object type \code{mle}.}
}
\description{
Computes the variance-covariance matrix of the MLE.
}
