% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{vcov.mle_fit}
\alias{vcov.mle_fit}
\title{Computes the variance-covariance matrix of `mle_fit` object.}
\usage{
\method{vcov}{mle_fit}(object, ...)
}
\arguments{
\item{object}{the `mle_fit` object to obtain the variance-covariance of}

\item{...}{additional arguments to pass (not used)}
}
\value{
the variance-covariance matrix
}
\description{
Computes the variance-covariance matrix of `mle_fit` object.
}
\examples{
fit <- mle(theta.hat = c(mu = 5, sigma2 = 4),
  sigma = diag(c(0.04, 0.32)), loglike = -120, nobs = 100L)
vcov(fit)
}
