% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weibull.R
\name{weibull_fim}
\alias{weibull_fim}
\title{weibull_fim}
\usage{
weibull_fim(x)
}
\arguments{
\item{x}{observed data}
}
\description{
Observed Fisher information generator given data `x` for the weibull
distribution
}
