% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_weibull.R
\name{weibull_score}
\alias{weibull_score}
\title{score function generator given data \code{x} for the weibull
distribution given a simple random sample.}
\usage{
weibull_score(x)
}
\arguments{
\item{x}{data}
}
\description{
score function generator given data \code{x} for the weibull
distribution given a simple random sample.
}
